@@ -1,27 +1,29 @@
 module github.com/koustech/mastermind
 
-go 1.18
+go 1.21
+
+toolchain go1.22.5
 
 require (
 	github.com/aler9/gomavlib v0.0.0-20220722052739-8f67720006ed
-	github.com/google/uuid v1.1.2
+	github.com/google/uuid v1.6.0
 	github.com/ispringtech/eventbus v0.0.0-20190701203509-53c525703608
 	github.com/stretchr/testify v1.8.0
 	go.uber.org/zap v1.21.0
-	google.golang.org/grpc v1.48.0
-	google.golang.org/protobuf v1.28.0
+	google.golang.org/grpc v1.65.0
+	google.golang.org/protobuf v1.34.1
 )
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/golang/protobuf v1.5.2 // indirect
+	github.com/golang/protobuf v1.5.4 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/tarm/serial v0.0.0-20180830185346-98f6abe2eb07 // indirect
 	go.uber.org/atomic v1.9.0 // indirect
 	go.uber.org/multierr v1.8.0 // indirect
-	golang.org/x/net v0.7.0 // indirect
-	golang.org/x/sys v0.5.0 // indirect
-	golang.org/x/text v0.7.0 // indirect
+	golang.org/x/net v0.26.0 // indirect
+	golang.org/x/sys v0.21.0 // indirect
+	golang.org/x/text v0.16.0 // indirect
 	google.golang.org/genproto v0.0.0-20200526211855-cb27e3aa2013 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
