@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.30.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: proto/approach/v1/approach.proto
 
@@ -84,13 +84,14 @@ type Plane struct {
 	Lon          float64      `protobuf:"fixed64,3,opt,name=lon,proto3" json:"lon,omitempty"`                                                                    // Plane's longitude in degrees
 	Alt          float64      `protobuf:"fixed64,4,opt,name=alt,proto3" json:"alt,omitempty"`                                                                    // Plane's altitude in metres
 	Heading      float64      `protobuf:"fixed64,5,opt,name=heading,proto3" json:"heading,omitempty"`                                                            // Plane's heading in degrees from North
-	Speed        float64      `protobuf:"fixed64,6,opt,name=speed,proto3" json:"speed,omitempty"`                                                                // Plane's groundspeend in m/s
+	Groundspeed  float64      `protobuf:"fixed64,6,opt,name=groundspeed,proto3" json:"groundspeed,omitempty"`                                                    // Plane's groundspeend in m/s
 	AgilityLevel AgilityLevel `protobuf:"varint,7,opt,name=agility_level,json=agilityLevel,proto3,enum=approach.v1.AgilityLevel" json:"agility_level,omitempty"` // Plane's agility level. High or 1 is most
 	// agile, low or 3 is least agile
 	Score float64 `protobuf:"fixed64,8,opt,name=score,proto3" json:"score,omitempty"` // Probability of locking on plane, between 0 and 100. 100
 	// is easiest to lock on
 	IsInFov        bool    `protobuf:"varint,9,opt,name=is_in_fov,json=isInFov,proto3" json:"is_in_fov,omitempty"`                      // Shows if the plane is in FOV angle
 	TargetDistance float64 `protobuf:"fixed64,10,opt,name=target_distance,json=targetDistance,proto3" json:"target_distance,omitempty"` //The distance of the approaching plane
+	Airspeed       float64 `protobuf:"fixed64,11,opt,name=airspeed,proto3" json:"airspeed,omitempty"`
 }
 
 func (x *Plane) Reset() {
@@ -160,9 +161,9 @@ func (x *Plane) GetHeading() float64 {
 	return 0
 }
 
-func (x *Plane) GetSpeed() float64 {
+func (x *Plane) GetGroundspeed() float64 {
 	if x != nil {
-		return x.Speed
+		return x.Groundspeed
 	}
 	return 0
 }
@@ -191,6 +192,13 @@ func (x *Plane) GetIsInFov() bool {
 func (x *Plane) GetTargetDistance() float64 {
 	if x != nil {
 		return x.TargetDistance
+	}
+	return 0
+}
+
+func (x *Plane) GetAirspeed() float64 {
+	if x != nil {
+		return x.Airspeed
 	}
 	return 0
 }
@@ -408,75 +416,77 @@ var file_proto_approach_v1_approach_proto_rawDesc = []byte{
 	0x0a, 0x20, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2f, 0x61, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68,
 	0x2f, 0x76, 0x31, 0x2f, 0x61, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x2e, 0x70, 0x72, 0x6f,
 	0x74, 0x6f, 0x12, 0x0b, 0x61, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x2e, 0x76, 0x31, 0x22,
-	0xa3, 0x02, 0x0a, 0x05, 0x50, 0x6c, 0x61, 0x6e, 0x65, 0x12, 0x19, 0x0a, 0x08, 0x70, 0x6c, 0x61,
+	0xcb, 0x02, 0x0a, 0x05, 0x50, 0x6c, 0x61, 0x6e, 0x65, 0x12, 0x19, 0x0a, 0x08, 0x70, 0x6c, 0x61,
 	0x6e, 0x65, 0x5f, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x05, 0x52, 0x07, 0x70, 0x6c, 0x61,
 	0x6e, 0x65, 0x49, 0x64, 0x12, 0x10, 0x0a, 0x03, 0x6c, 0x61, 0x74, 0x18, 0x02, 0x20, 0x01, 0x28,
 	0x01, 0x52, 0x03, 0x6c, 0x61, 0x74, 0x12, 0x10, 0x0a, 0x03, 0x6c, 0x6f, 0x6e, 0x18, 0x03, 0x20,
 	0x01, 0x28, 0x01, 0x52, 0x03, 0x6c, 0x6f, 0x6e, 0x12, 0x10, 0x0a, 0x03, 0x61, 0x6c, 0x74, 0x18,
 	0x04, 0x20, 0x01, 0x28, 0x01, 0x52, 0x03, 0x61, 0x6c, 0x74, 0x12, 0x18, 0x0a, 0x07, 0x68, 0x65,
 	0x61, 0x64, 0x69, 0x6e, 0x67, 0x18, 0x05, 0x20, 0x01, 0x28, 0x01, 0x52, 0x07, 0x68, 0x65, 0x61,
-	0x64, 0x69, 0x6e, 0x67, 0x12, 0x14, 0x0a, 0x05, 0x73, 0x70, 0x65, 0x65, 0x64, 0x18, 0x06, 0x20,
-	0x01, 0x28, 0x01, 0x52, 0x05, 0x73, 0x70, 0x65, 0x65, 0x64, 0x12, 0x3e, 0x0a, 0x0d, 0x61, 0x67,
-	0x69, 0x6c, 0x69, 0x74, 0x79, 0x5f, 0x6c, 0x65, 0x76, 0x65, 0x6c, 0x18, 0x07, 0x20, 0x01, 0x28,
-	0x0e, 0x32, 0x19, 0x2e, 0x61, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x2e, 0x76, 0x31, 0x2e,
-	0x41, 0x67, 0x69, 0x6c, 0x69, 0x74, 0x79, 0x4c, 0x65, 0x76, 0x65, 0x6c, 0x52, 0x0c, 0x61, 0x67,
-	0x69, 0x6c, 0x69, 0x74, 0x79, 0x4c, 0x65, 0x76, 0x65, 0x6c, 0x12, 0x14, 0x0a, 0x05, 0x73, 0x63,
-	0x6f, 0x72, 0x65, 0x18, 0x08, 0x20, 0x01, 0x28, 0x01, 0x52, 0x05, 0x73, 0x63, 0x6f, 0x72, 0x65,
-	0x12, 0x1a, 0x0a, 0x09, 0x69, 0x73, 0x5f, 0x69, 0x6e, 0x5f, 0x66, 0x6f, 0x76, 0x18, 0x09, 0x20,
-	0x01, 0x28, 0x08, 0x52, 0x07, 0x69, 0x73, 0x49, 0x6e, 0x46, 0x6f, 0x76, 0x12, 0x27, 0x0a, 0x0f,
-	0x74, 0x61, 0x72, 0x67, 0x65, 0x74, 0x5f, 0x64, 0x69, 0x73, 0x74, 0x61, 0x6e, 0x63, 0x65, 0x18,
-	0x0a, 0x20, 0x01, 0x28, 0x01, 0x52, 0x0e, 0x74, 0x61, 0x72, 0x67, 0x65, 0x74, 0x44, 0x69, 0x73,
-	0x74, 0x61, 0x6e, 0x63, 0x65, 0x22, 0x12, 0x0a, 0x10, 0x47, 0x65, 0x74, 0x50, 0x6c, 0x61, 0x6e,
-	0x65, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x22, 0x3f, 0x0a, 0x11, 0x47, 0x65, 0x74,
-	0x50, 0x6c, 0x61, 0x6e, 0x65, 0x73, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x2a,
-	0x0a, 0x06, 0x70, 0x6c, 0x61, 0x6e, 0x65, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x12,
-	0x2e, 0x61, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x2e, 0x76, 0x31, 0x2e, 0x50, 0x6c, 0x61,
-	0x6e, 0x65, 0x52, 0x06, 0x70, 0x6c, 0x61, 0x6e, 0x65, 0x73, 0x22, 0x7e, 0x0a, 0x14, 0x41, 0x70,
-	0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x50, 0x6c, 0x61, 0x6e, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65,
-	0x73, 0x74, 0x12, 0x19, 0x0a, 0x08, 0x70, 0x6c, 0x61, 0x6e, 0x65, 0x5f, 0x69, 0x64, 0x18, 0x01,
-	0x20, 0x01, 0x28, 0x05, 0x52, 0x07, 0x70, 0x6c, 0x61, 0x6e, 0x65, 0x49, 0x64, 0x12, 0x23, 0x0a,
-	0x0d, 0x6c, 0x6f, 0x63, 0x6b, 0x5f, 0x64, 0x69, 0x73, 0x74, 0x61, 0x6e, 0x63, 0x65, 0x18, 0x02,
-	0x20, 0x01, 0x28, 0x05, 0x52, 0x0c, 0x6c, 0x6f, 0x63, 0x6b, 0x44, 0x69, 0x73, 0x74, 0x61, 0x6e,
-	0x63, 0x65, 0x12, 0x26, 0x0a, 0x0f, 0x77, 0x69, 0x6c, 0x6c, 0x5f, 0x64, 0x65, 0x74, 0x65, 0x63,
-	0x74, 0x5f, 0x67, 0x70, 0x73, 0x18, 0x03, 0x20, 0x01, 0x28, 0x08, 0x52, 0x0d, 0x77, 0x69, 0x6c,
-	0x6c, 0x44, 0x65, 0x74, 0x65, 0x63, 0x74, 0x47, 0x70, 0x73, 0x22, 0x56, 0x0a, 0x15, 0x41, 0x70,
-	0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x50, 0x6c, 0x61, 0x6e, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f,
-	0x6e, 0x73, 0x65, 0x12, 0x2b, 0x0a, 0x11, 0x63, 0x6f, 0x6d, 0x6d, 0x61, 0x6e, 0x64, 0x5f, 0x73,
-	0x75, 0x63, 0x63, 0x65, 0x65, 0x64, 0x65, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x08, 0x52, 0x10,
-	0x63, 0x6f, 0x6d, 0x6d, 0x61, 0x6e, 0x64, 0x53, 0x75, 0x63, 0x63, 0x65, 0x65, 0x64, 0x65, 0x64,
-	0x12, 0x10, 0x0a, 0x03, 0x6d, 0x73, 0x67, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x6d,
-	0x73, 0x67, 0x2a, 0x76, 0x0a, 0x0c, 0x41, 0x67, 0x69, 0x6c, 0x69, 0x74, 0x79, 0x4c, 0x65, 0x76,
-	0x65, 0x6c, 0x12, 0x1d, 0x0a, 0x19, 0x41, 0x47, 0x49, 0x4c, 0x49, 0x54, 0x59, 0x5f, 0x4c, 0x45,
-	0x56, 0x45, 0x4c, 0x5f, 0x55, 0x4e, 0x53, 0x50, 0x45, 0x43, 0x49, 0x46, 0x49, 0x45, 0x44, 0x10,
-	0x00, 0x12, 0x16, 0x0a, 0x12, 0x41, 0x47, 0x49, 0x4c, 0x49, 0x54, 0x59, 0x5f, 0x4c, 0x45, 0x56,
-	0x45, 0x4c, 0x5f, 0x48, 0x49, 0x47, 0x48, 0x10, 0x01, 0x12, 0x18, 0x0a, 0x14, 0x41, 0x47, 0x49,
-	0x4c, 0x49, 0x54, 0x59, 0x5f, 0x4c, 0x45, 0x56, 0x45, 0x4c, 0x5f, 0x4d, 0x45, 0x44, 0x49, 0x55,
-	0x4d, 0x10, 0x02, 0x12, 0x15, 0x0a, 0x11, 0x41, 0x47, 0x49, 0x4c, 0x49, 0x54, 0x59, 0x5f, 0x4c,
-	0x45, 0x56, 0x45, 0x4c, 0x5f, 0x4c, 0x4f, 0x57, 0x10, 0x03, 0x32, 0xb7, 0x01, 0x0a, 0x0f, 0x41,
-	0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x53, 0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x12, 0x4c,
-	0x0a, 0x09, 0x47, 0x65, 0x74, 0x50, 0x6c, 0x61, 0x6e, 0x65, 0x73, 0x12, 0x1d, 0x2e, 0x61, 0x70,
-	0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x2e, 0x76, 0x31, 0x2e, 0x47, 0x65, 0x74, 0x50, 0x6c, 0x61,
-	0x6e, 0x65, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1e, 0x2e, 0x61, 0x70, 0x70,
-	0x72, 0x6f, 0x61, 0x63, 0x68, 0x2e, 0x76, 0x31, 0x2e, 0x47, 0x65, 0x74, 0x50, 0x6c, 0x61, 0x6e,
-	0x65, 0x73, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x30, 0x01, 0x12, 0x56, 0x0a, 0x0d,
-	0x41, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x50, 0x6c, 0x61, 0x6e, 0x65, 0x12, 0x21, 0x2e,
-	0x61, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x2e, 0x76, 0x31, 0x2e, 0x41, 0x70, 0x70, 0x72,
-	0x6f, 0x61, 0x63, 0x68, 0x50, 0x6c, 0x61, 0x6e, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74,
-	0x1a, 0x22, 0x2e, 0x61, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x2e, 0x76, 0x31, 0x2e, 0x41,
-	0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x50, 0x6c, 0x61, 0x6e, 0x65, 0x52, 0x65, 0x73, 0x70,
-	0x6f, 0x6e, 0x73, 0x65, 0x42, 0xb7, 0x01, 0x0a, 0x0f, 0x63, 0x6f, 0x6d, 0x2e, 0x61, 0x70, 0x70,
-	0x72, 0x6f, 0x61, 0x63, 0x68, 0x2e, 0x76, 0x31, 0x42, 0x0d, 0x41, 0x70, 0x70, 0x72, 0x6f, 0x61,
-	0x63, 0x68, 0x50, 0x72, 0x6f, 0x74, 0x6f, 0x50, 0x01, 0x5a, 0x48, 0x67, 0x69, 0x74, 0x68, 0x75,
-	0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x6b, 0x6f, 0x75, 0x73, 0x74, 0x65, 0x63, 0x68, 0x2f, 0x6d,
-	0x61, 0x73, 0x74, 0x65, 0x72, 0x6d, 0x69, 0x6e, 0x64, 0x2f, 0x67, 0x65, 0x6e, 0x2f, 0x70, 0x72,
-	0x6f, 0x74, 0x6f, 0x2f, 0x67, 0x6f, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2f, 0x61, 0x70, 0x70,
-	0x72, 0x6f, 0x61, 0x63, 0x68, 0x2f, 0x76, 0x31, 0x3b, 0x61, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63,
-	0x68, 0x76, 0x31, 0xa2, 0x02, 0x03, 0x41, 0x58, 0x58, 0xaa, 0x02, 0x0b, 0x41, 0x70, 0x70, 0x72,
-	0x6f, 0x61, 0x63, 0x68, 0x2e, 0x56, 0x31, 0xca, 0x02, 0x0b, 0x41, 0x70, 0x70, 0x72, 0x6f, 0x61,
-	0x63, 0x68, 0x5c, 0x56, 0x31, 0xe2, 0x02, 0x17, 0x41, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68,
-	0x5c, 0x56, 0x31, 0x5c, 0x47, 0x50, 0x42, 0x4d, 0x65, 0x74, 0x61, 0x64, 0x61, 0x74, 0x61, 0xea,
-	0x02, 0x0c, 0x41, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x3a, 0x3a, 0x56, 0x31, 0x62, 0x06,
-	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
+	0x64, 0x69, 0x6e, 0x67, 0x12, 0x20, 0x0a, 0x0b, 0x67, 0x72, 0x6f, 0x75, 0x6e, 0x64, 0x73, 0x70,
+	0x65, 0x65, 0x64, 0x18, 0x06, 0x20, 0x01, 0x28, 0x01, 0x52, 0x0b, 0x67, 0x72, 0x6f, 0x75, 0x6e,
+	0x64, 0x73, 0x70, 0x65, 0x65, 0x64, 0x12, 0x3e, 0x0a, 0x0d, 0x61, 0x67, 0x69, 0x6c, 0x69, 0x74,
+	0x79, 0x5f, 0x6c, 0x65, 0x76, 0x65, 0x6c, 0x18, 0x07, 0x20, 0x01, 0x28, 0x0e, 0x32, 0x19, 0x2e,
+	0x61, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x2e, 0x76, 0x31, 0x2e, 0x41, 0x67, 0x69, 0x6c,
+	0x69, 0x74, 0x79, 0x4c, 0x65, 0x76, 0x65, 0x6c, 0x52, 0x0c, 0x61, 0x67, 0x69, 0x6c, 0x69, 0x74,
+	0x79, 0x4c, 0x65, 0x76, 0x65, 0x6c, 0x12, 0x14, 0x0a, 0x05, 0x73, 0x63, 0x6f, 0x72, 0x65, 0x18,
+	0x08, 0x20, 0x01, 0x28, 0x01, 0x52, 0x05, 0x73, 0x63, 0x6f, 0x72, 0x65, 0x12, 0x1a, 0x0a, 0x09,
+	0x69, 0x73, 0x5f, 0x69, 0x6e, 0x5f, 0x66, 0x6f, 0x76, 0x18, 0x09, 0x20, 0x01, 0x28, 0x08, 0x52,
+	0x07, 0x69, 0x73, 0x49, 0x6e, 0x46, 0x6f, 0x76, 0x12, 0x27, 0x0a, 0x0f, 0x74, 0x61, 0x72, 0x67,
+	0x65, 0x74, 0x5f, 0x64, 0x69, 0x73, 0x74, 0x61, 0x6e, 0x63, 0x65, 0x18, 0x0a, 0x20, 0x01, 0x28,
+	0x01, 0x52, 0x0e, 0x74, 0x61, 0x72, 0x67, 0x65, 0x74, 0x44, 0x69, 0x73, 0x74, 0x61, 0x6e, 0x63,
+	0x65, 0x12, 0x1a, 0x0a, 0x08, 0x61, 0x69, 0x72, 0x73, 0x70, 0x65, 0x65, 0x64, 0x18, 0x0b, 0x20,
+	0x01, 0x28, 0x01, 0x52, 0x08, 0x61, 0x69, 0x72, 0x73, 0x70, 0x65, 0x65, 0x64, 0x22, 0x12, 0x0a,
+	0x10, 0x47, 0x65, 0x74, 0x50, 0x6c, 0x61, 0x6e, 0x65, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73,
+	0x74, 0x22, 0x3f, 0x0a, 0x11, 0x47, 0x65, 0x74, 0x50, 0x6c, 0x61, 0x6e, 0x65, 0x73, 0x52, 0x65,
+	0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x2a, 0x0a, 0x06, 0x70, 0x6c, 0x61, 0x6e, 0x65, 0x73,
+	0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x12, 0x2e, 0x61, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63,
+	0x68, 0x2e, 0x76, 0x31, 0x2e, 0x50, 0x6c, 0x61, 0x6e, 0x65, 0x52, 0x06, 0x70, 0x6c, 0x61, 0x6e,
+	0x65, 0x73, 0x22, 0x7e, 0x0a, 0x14, 0x41, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x50, 0x6c,
+	0x61, 0x6e, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x19, 0x0a, 0x08, 0x70, 0x6c,
+	0x61, 0x6e, 0x65, 0x5f, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x05, 0x52, 0x07, 0x70, 0x6c,
+	0x61, 0x6e, 0x65, 0x49, 0x64, 0x12, 0x23, 0x0a, 0x0d, 0x6c, 0x6f, 0x63, 0x6b, 0x5f, 0x64, 0x69,
+	0x73, 0x74, 0x61, 0x6e, 0x63, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x05, 0x52, 0x0c, 0x6c, 0x6f,
+	0x63, 0x6b, 0x44, 0x69, 0x73, 0x74, 0x61, 0x6e, 0x63, 0x65, 0x12, 0x26, 0x0a, 0x0f, 0x77, 0x69,
+	0x6c, 0x6c, 0x5f, 0x64, 0x65, 0x74, 0x65, 0x63, 0x74, 0x5f, 0x67, 0x70, 0x73, 0x18, 0x03, 0x20,
+	0x01, 0x28, 0x08, 0x52, 0x0d, 0x77, 0x69, 0x6c, 0x6c, 0x44, 0x65, 0x74, 0x65, 0x63, 0x74, 0x47,
+	0x70, 0x73, 0x22, 0x56, 0x0a, 0x15, 0x41, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x50, 0x6c,
+	0x61, 0x6e, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x2b, 0x0a, 0x11, 0x63,
+	0x6f, 0x6d, 0x6d, 0x61, 0x6e, 0x64, 0x5f, 0x73, 0x75, 0x63, 0x63, 0x65, 0x65, 0x64, 0x65, 0x64,
+	0x18, 0x01, 0x20, 0x01, 0x28, 0x08, 0x52, 0x10, 0x63, 0x6f, 0x6d, 0x6d, 0x61, 0x6e, 0x64, 0x53,
+	0x75, 0x63, 0x63, 0x65, 0x65, 0x64, 0x65, 0x64, 0x12, 0x10, 0x0a, 0x03, 0x6d, 0x73, 0x67, 0x18,
+	0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x6d, 0x73, 0x67, 0x2a, 0x76, 0x0a, 0x0c, 0x41, 0x67,
+	0x69, 0x6c, 0x69, 0x74, 0x79, 0x4c, 0x65, 0x76, 0x65, 0x6c, 0x12, 0x1d, 0x0a, 0x19, 0x41, 0x47,
+	0x49, 0x4c, 0x49, 0x54, 0x59, 0x5f, 0x4c, 0x45, 0x56, 0x45, 0x4c, 0x5f, 0x55, 0x4e, 0x53, 0x50,
+	0x45, 0x43, 0x49, 0x46, 0x49, 0x45, 0x44, 0x10, 0x00, 0x12, 0x16, 0x0a, 0x12, 0x41, 0x47, 0x49,
+	0x4c, 0x49, 0x54, 0x59, 0x5f, 0x4c, 0x45, 0x56, 0x45, 0x4c, 0x5f, 0x48, 0x49, 0x47, 0x48, 0x10,
+	0x01, 0x12, 0x18, 0x0a, 0x14, 0x41, 0x47, 0x49, 0x4c, 0x49, 0x54, 0x59, 0x5f, 0x4c, 0x45, 0x56,
+	0x45, 0x4c, 0x5f, 0x4d, 0x45, 0x44, 0x49, 0x55, 0x4d, 0x10, 0x02, 0x12, 0x15, 0x0a, 0x11, 0x41,
+	0x47, 0x49, 0x4c, 0x49, 0x54, 0x59, 0x5f, 0x4c, 0x45, 0x56, 0x45, 0x4c, 0x5f, 0x4c, 0x4f, 0x57,
+	0x10, 0x03, 0x32, 0xb7, 0x01, 0x0a, 0x0f, 0x41, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x53,
+	0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x12, 0x4c, 0x0a, 0x09, 0x47, 0x65, 0x74, 0x50, 0x6c, 0x61,
+	0x6e, 0x65, 0x73, 0x12, 0x1d, 0x2e, 0x61, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x2e, 0x76,
+	0x31, 0x2e, 0x47, 0x65, 0x74, 0x50, 0x6c, 0x61, 0x6e, 0x65, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65,
+	0x73, 0x74, 0x1a, 0x1e, 0x2e, 0x61, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x2e, 0x76, 0x31,
+	0x2e, 0x47, 0x65, 0x74, 0x50, 0x6c, 0x61, 0x6e, 0x65, 0x73, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e,
+	0x73, 0x65, 0x30, 0x01, 0x12, 0x56, 0x0a, 0x0d, 0x41, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68,
+	0x50, 0x6c, 0x61, 0x6e, 0x65, 0x12, 0x21, 0x2e, 0x61, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68,
+	0x2e, 0x76, 0x31, 0x2e, 0x41, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x50, 0x6c, 0x61, 0x6e,
+	0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x22, 0x2e, 0x61, 0x70, 0x70, 0x72, 0x6f,
+	0x61, 0x63, 0x68, 0x2e, 0x76, 0x31, 0x2e, 0x41, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x50,
+	0x6c, 0x61, 0x6e, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x42, 0xb7, 0x01, 0x0a,
+	0x0f, 0x63, 0x6f, 0x6d, 0x2e, 0x61, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x2e, 0x76, 0x31,
+	0x42, 0x0d, 0x41, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x50, 0x72, 0x6f, 0x74, 0x6f, 0x50,
+	0x01, 0x5a, 0x48, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x6b, 0x6f,
+	0x75, 0x73, 0x74, 0x65, 0x63, 0x68, 0x2f, 0x6d, 0x61, 0x73, 0x74, 0x65, 0x72, 0x6d, 0x69, 0x6e,
+	0x64, 0x2f, 0x67, 0x65, 0x6e, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2f, 0x67, 0x6f, 0x2f, 0x70,
+	0x72, 0x6f, 0x74, 0x6f, 0x2f, 0x61, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x2f, 0x76, 0x31,
+	0x3b, 0x61, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x76, 0x31, 0xa2, 0x02, 0x03, 0x41, 0x58,
+	0x58, 0xaa, 0x02, 0x0b, 0x41, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x2e, 0x56, 0x31, 0xca,
+	0x02, 0x0b, 0x41, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x5c, 0x56, 0x31, 0xe2, 0x02, 0x17,
+	0x41, 0x70, 0x70, 0x72, 0x6f, 0x61, 0x63, 0x68, 0x5c, 0x56, 0x31, 0x5c, 0x47, 0x50, 0x42, 0x4d,
+	0x65, 0x74, 0x61, 0x64, 0x61, 0x74, 0x61, 0xea, 0x02, 0x0c, 0x41, 0x70, 0x70, 0x72, 0x6f, 0x61,
+	0x63, 0x68, 0x3a, 0x3a, 0x56, 0x31, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
 }
 
 var (
@@ -493,7 +503,7 @@ func file_proto_approach_v1_approach_proto_rawDescGZIP() []byte {
 
 var file_proto_approach_v1_approach_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_proto_approach_v1_approach_proto_msgTypes = make([]protoimpl.MessageInfo, 5)
-var file_proto_approach_v1_approach_proto_goTypes = []interface{}{
+var file_proto_approach_v1_approach_proto_goTypes = []any{
 	(AgilityLevel)(0),             // 0: approach.v1.AgilityLevel
 	(*Plane)(nil),                 // 1: approach.v1.Plane
 	(*GetPlanesRequest)(nil),      // 2: approach.v1.GetPlanesRequest
@@ -521,7 +531,7 @@ func file_proto_approach_v1_approach_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_proto_approach_v1_approach_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_approach_v1_approach_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Plane); i {
 			case 0:
 				return &v.state
@@ -533,7 +543,7 @@ func file_proto_approach_v1_approach_proto_init() {
 				return nil
 			}
 		}
-		file_proto_approach_v1_approach_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_approach_v1_approach_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*GetPlanesRequest); i {
 			case 0:
 				return &v.state
@@ -545,7 +555,7 @@ func file_proto_approach_v1_approach_proto_init() {
 				return nil
 			}
 		}
-		file_proto_approach_v1_approach_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_approach_v1_approach_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*GetPlanesResponse); i {
 			case 0:
 				return &v.state
@@ -557,7 +567,7 @@ func file_proto_approach_v1_approach_proto_init() {
 				return nil
 			}
 		}
-		file_proto_approach_v1_approach_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_approach_v1_approach_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*ApproachPlaneRequest); i {
 			case 0:
 				return &v.state
@@ -569,7 +579,7 @@ func file_proto_approach_v1_approach_proto_init() {
 				return nil
 			}
 		}
-		file_proto_approach_v1_approach_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_approach_v1_approach_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*ApproachPlaneResponse); i {
 			case 0:
 				return &v.state
