@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.30.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: proto/mastermind/v1/mastermind.proto
 
@@ -1510,7 +1510,7 @@ func file_proto_mastermind_v1_mastermind_proto_rawDescGZIP() []byte {
 
 var file_proto_mastermind_v1_mastermind_proto_enumTypes = make([]protoimpl.EnumInfo, 3)
 var file_proto_mastermind_v1_mastermind_proto_msgTypes = make([]protoimpl.MessageInfo, 16)
-var file_proto_mastermind_v1_mastermind_proto_goTypes = []interface{}{
+var file_proto_mastermind_v1_mastermind_proto_goTypes = []any{
 	(MissionState)(0),                    // 0: mastermind.v1.MissionState
 	(StateTransition)(0),                 // 1: mastermind.v1.StateTransition
 	(PIDLevel)(0),                        // 2: mastermind.v1.PIDLevel
@@ -1566,7 +1566,7 @@ func file_proto_mastermind_v1_mastermind_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_proto_mastermind_v1_mastermind_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_mastermind_v1_mastermind_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdateStateRequest); i {
 			case 0:
 				return &v.state
@@ -1578,7 +1578,7 @@ func file_proto_mastermind_v1_mastermind_proto_init() {
 				return nil
 			}
 		}
-		file_proto_mastermind_v1_mastermind_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_mastermind_v1_mastermind_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*UpdateStateResponse); i {
 			case 0:
 				return &v.state
@@ -1590,7 +1590,7 @@ func file_proto_mastermind_v1_mastermind_proto_init() {
 				return nil
 			}
 		}
-		file_proto_mastermind_v1_mastermind_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_mastermind_v1_mastermind_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*GetTelemetryRequest); i {
 			case 0:
 				return &v.state
@@ -1602,7 +1602,7 @@ func file_proto_mastermind_v1_mastermind_proto_init() {
 				return nil
 			}
 		}
-		file_proto_mastermind_v1_mastermind_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_mastermind_v1_mastermind_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*GetTelemetryResponse); i {
 			case 0:
 				return &v.state
@@ -1614,7 +1614,7 @@ func file_proto_mastermind_v1_mastermind_proto_init() {
 				return nil
 			}
 		}
-		file_proto_mastermind_v1_mastermind_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_mastermind_v1_mastermind_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*Target); i {
 			case 0:
 				return &v.state
@@ -1626,7 +1626,7 @@ func file_proto_mastermind_v1_mastermind_proto_init() {
 				return nil
 			}
 		}
-		file_proto_mastermind_v1_mastermind_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_mastermind_v1_mastermind_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*GetDetailedTelemetryRequest); i {
 			case 0:
 				return &v.state
@@ -1638,7 +1638,7 @@ func file_proto_mastermind_v1_mastermind_proto_init() {
 				return nil
 			}
 		}
-		file_proto_mastermind_v1_mastermind_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_mastermind_v1_mastermind_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*GetDetailedTelemetryResponse); i {
 			case 0:
 				return &v.state
@@ -1650,7 +1650,7 @@ func file_proto_mastermind_v1_mastermind_proto_init() {
 				return nil
 			}
 		}
-		file_proto_mastermind_v1_mastermind_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_mastermind_v1_mastermind_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*SetPIDLevelRequest); i {
 			case 0:
 				return &v.state
@@ -1662,7 +1662,7 @@ func file_proto_mastermind_v1_mastermind_proto_init() {
 				return nil
 			}
 		}
-		file_proto_mastermind_v1_mastermind_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_mastermind_v1_mastermind_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*SetPIDLevelResponse); i {
 			case 0:
 				return &v.state
@@ -1674,7 +1674,7 @@ func file_proto_mastermind_v1_mastermind_proto_init() {
 				return nil
 			}
 		}
-		file_proto_mastermind_v1_mastermind_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_mastermind_v1_mastermind_proto_msgTypes[9].Exporter = func(v any, i int) any {
 			switch v := v.(*SetSpeedRequest); i {
 			case 0:
 				return &v.state
@@ -1686,7 +1686,7 @@ func file_proto_mastermind_v1_mastermind_proto_init() {
 				return nil
 			}
 		}
-		file_proto_mastermind_v1_mastermind_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_mastermind_v1_mastermind_proto_msgTypes[10].Exporter = func(v any, i int) any {
 			switch v := v.(*SetSpeedResponse); i {
 			case 0:
 				return &v.state
@@ -1698,7 +1698,7 @@ func file_proto_mastermind_v1_mastermind_proto_init() {
 				return nil
 			}
 		}
-		file_proto_mastermind_v1_mastermind_proto_msgTypes[11].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_mastermind_v1_mastermind_proto_msgTypes[11].Exporter = func(v any, i int) any {
 			switch v := v.(*GotoWaypointRequest); i {
 			case 0:
 				return &v.state
@@ -1710,7 +1710,7 @@ func file_proto_mastermind_v1_mastermind_proto_init() {
 				return nil
 			}
 		}
-		file_proto_mastermind_v1_mastermind_proto_msgTypes[12].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_mastermind_v1_mastermind_proto_msgTypes[12].Exporter = func(v any, i int) any {
 			switch v := v.(*SetAttitudeRequest); i {
 			case 0:
 				return &v.state
@@ -1722,7 +1722,7 @@ func file_proto_mastermind_v1_mastermind_proto_init() {
 				return nil
 			}
 		}
-		file_proto_mastermind_v1_mastermind_proto_msgTypes[13].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_mastermind_v1_mastermind_proto_msgTypes[13].Exporter = func(v any, i int) any {
 			switch v := v.(*SetAttitudeResponse); i {
 			case 0:
 				return &v.state
@@ -1734,7 +1734,7 @@ func file_proto_mastermind_v1_mastermind_proto_init() {
 				return nil
 			}
 		}
-		file_proto_mastermind_v1_mastermind_proto_msgTypes[14].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_mastermind_v1_mastermind_proto_msgTypes[14].Exporter = func(v any, i int) any {
 			switch v := v.(*GotoWaypointResponse); i {
 			case 0:
 				return &v.state
@@ -1746,7 +1746,7 @@ func file_proto_mastermind_v1_mastermind_proto_init() {
 				return nil
 			}
 		}
-		file_proto_mastermind_v1_mastermind_proto_msgTypes[15].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_mastermind_v1_mastermind_proto_msgTypes[15].Exporter = func(v any, i int) any {
 			switch v := v.(*Time); i {
 			case 0:
 				return &v.state
@@ -1759,7 +1759,7 @@ func file_proto_mastermind_v1_mastermind_proto_init() {
 			}
 		}
 	}
-	file_proto_mastermind_v1_mastermind_proto_msgTypes[6].OneofWrappers = []interface{}{}
+	file_proto_mastermind_v1_mastermind_proto_msgTypes[6].OneofWrappers = []any{}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
