@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.30.0
+// 	protoc-gen-go v1.34.2
 // 	protoc        (unknown)
 // source: proto/proxy/v1/proxy.proto
 
@@ -945,6 +945,154 @@ func (*SendKamikazePacketResponse) Descriptor() ([]byte, []int) {
 	return file_proto_proxy_v1_proxy_proto_rawDescGZIP(), []int{12}
 }
 
+type RedZonePacket struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Lat float32 `protobuf:"fixed32,1,opt,name=lat,proto3" json:"lat,omitempty"`
+	Lon float32 `protobuf:"fixed32,2,opt,name=lon,proto3" json:"lon,omitempty"`
+	Rad float32 `protobuf:"fixed32,3,opt,name=rad,proto3" json:"rad,omitempty"`
+}
+
+func (x *RedZonePacket) Reset() {
+	*x = RedZonePacket{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_proto_proxy_v1_proxy_proto_msgTypes[13]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *RedZonePacket) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*RedZonePacket) ProtoMessage() {}
+
+func (x *RedZonePacket) ProtoReflect() protoreflect.Message {
+	mi := &file_proto_proxy_v1_proxy_proto_msgTypes[13]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use RedZonePacket.ProtoReflect.Descriptor instead.
+func (*RedZonePacket) Descriptor() ([]byte, []int) {
+	return file_proto_proxy_v1_proxy_proto_rawDescGZIP(), []int{13}
+}
+
+func (x *RedZonePacket) GetLat() float32 {
+	if x != nil {
+		return x.Lat
+	}
+	return 0
+}
+
+func (x *RedZonePacket) GetLon() float32 {
+	if x != nil {
+		return x.Lon
+	}
+	return 0
+}
+
+func (x *RedZonePacket) GetRad() float32 {
+	if x != nil {
+		return x.Rad
+	}
+	return 0
+}
+
+type GetRedZonePacketRequest struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+}
+
+func (x *GetRedZonePacketRequest) Reset() {
+	*x = GetRedZonePacketRequest{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_proto_proxy_v1_proxy_proto_msgTypes[14]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *GetRedZonePacketRequest) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*GetRedZonePacketRequest) ProtoMessage() {}
+
+func (x *GetRedZonePacketRequest) ProtoReflect() protoreflect.Message {
+	mi := &file_proto_proxy_v1_proxy_proto_msgTypes[14]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use GetRedZonePacketRequest.ProtoReflect.Descriptor instead.
+func (*GetRedZonePacketRequest) Descriptor() ([]byte, []int) {
+	return file_proto_proxy_v1_proxy_proto_rawDescGZIP(), []int{14}
+}
+
+type GetRedZonePacketResponse struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	RedZones []*RedZonePacket `protobuf:"bytes,1,rep,name=redZones,proto3" json:"redZones,omitempty"`
+}
+
+func (x *GetRedZonePacketResponse) Reset() {
+	*x = GetRedZonePacketResponse{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_proto_proxy_v1_proxy_proto_msgTypes[15]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *GetRedZonePacketResponse) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*GetRedZonePacketResponse) ProtoMessage() {}
+
+func (x *GetRedZonePacketResponse) ProtoReflect() protoreflect.Message {
+	mi := &file_proto_proxy_v1_proxy_proto_msgTypes[15]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use GetRedZonePacketResponse.ProtoReflect.Descriptor instead.
+func (*GetRedZonePacketResponse) Descriptor() ([]byte, []int) {
+	return file_proto_proxy_v1_proxy_proto_rawDescGZIP(), []int{15}
+}
+
+func (x *GetRedZonePacketResponse) GetRedZones() []*RedZonePacket {
+	if x != nil {
+		return x.RedZones
+	}
+	return nil
+}
+
 var File_proto_proxy_v1_proxy_proto protoreflect.FileDescriptor
 
 var file_proto_proxy_v1_proxy_proto_rawDesc = []byte{
@@ -1066,41 +1214,58 @@ var file_proto_proxy_v1_proxy_proto_rawDesc = []byte{
 	0x71, 0x72, 0x5f, 0x74, 0x65, 0x78, 0x74, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x06, 0x71,
 	0x72, 0x54, 0x65, 0x78, 0x74, 0x22, 0x1c, 0x0a, 0x1a, 0x53, 0x65, 0x6e, 0x64, 0x4b, 0x61, 0x6d,
 	0x69, 0x6b, 0x61, 0x7a, 0x65, 0x50, 0x61, 0x63, 0x6b, 0x65, 0x74, 0x52, 0x65, 0x73, 0x70, 0x6f,
-	0x6e, 0x73, 0x65, 0x32, 0xf5, 0x02, 0x0a, 0x0c, 0x50, 0x72, 0x6f, 0x78, 0x79, 0x53, 0x65, 0x72,
-	0x76, 0x69, 0x63, 0x65, 0x12, 0x4f, 0x0a, 0x0c, 0x47, 0x65, 0x74, 0x54, 0x65, 0x6c, 0x65, 0x6d,
-	0x65, 0x74, 0x72, 0x79, 0x12, 0x1d, 0x2e, 0x70, 0x72, 0x6f, 0x78, 0x79, 0x2e, 0x76, 0x31, 0x2e,
-	0x47, 0x65, 0x74, 0x54, 0x65, 0x6c, 0x65, 0x6d, 0x65, 0x74, 0x72, 0x79, 0x52, 0x65, 0x71, 0x75,
-	0x65, 0x73, 0x74, 0x1a, 0x1e, 0x2e, 0x70, 0x72, 0x6f, 0x78, 0x79, 0x2e, 0x76, 0x31, 0x2e, 0x47,
-	0x65, 0x74, 0x54, 0x65, 0x6c, 0x65, 0x6d, 0x65, 0x74, 0x72, 0x79, 0x52, 0x65, 0x73, 0x70, 0x6f,
-	0x6e, 0x73, 0x65, 0x30, 0x01, 0x12, 0x5e, 0x0a, 0x11, 0x47, 0x65, 0x74, 0x45, 0x6e, 0x65, 0x6d,
-	0x79, 0x54, 0x65, 0x6c, 0x65, 0x6d, 0x65, 0x74, 0x72, 0x79, 0x12, 0x22, 0x2e, 0x70, 0x72, 0x6f,
-	0x78, 0x79, 0x2e, 0x76, 0x31, 0x2e, 0x47, 0x65, 0x74, 0x45, 0x6e, 0x65, 0x6d, 0x79, 0x54, 0x65,
-	0x6c, 0x65, 0x6d, 0x65, 0x74, 0x72, 0x79, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x23,
-	0x2e, 0x70, 0x72, 0x6f, 0x78, 0x79, 0x2e, 0x76, 0x31, 0x2e, 0x47, 0x65, 0x74, 0x45, 0x6e, 0x65,
-	0x6d, 0x79, 0x54, 0x65, 0x6c, 0x65, 0x6d, 0x65, 0x74, 0x72, 0x79, 0x52, 0x65, 0x73, 0x70, 0x6f,
-	0x6e, 0x73, 0x65, 0x30, 0x01, 0x12, 0x53, 0x0a, 0x0e, 0x53, 0x65, 0x6e, 0x64, 0x4c, 0x6f, 0x63,
-	0x6b, 0x50, 0x61, 0x63, 0x6b, 0x65, 0x74, 0x12, 0x1f, 0x2e, 0x70, 0x72, 0x6f, 0x78, 0x79, 0x2e,
-	0x76, 0x31, 0x2e, 0x53, 0x65, 0x6e, 0x64, 0x4c, 0x6f, 0x63, 0x6b, 0x50, 0x61, 0x63, 0x6b, 0x65,
-	0x74, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x20, 0x2e, 0x70, 0x72, 0x6f, 0x78, 0x79,
-	0x2e, 0x76, 0x31, 0x2e, 0x53, 0x65, 0x6e, 0x64, 0x4c, 0x6f, 0x63, 0x6b, 0x50, 0x61, 0x63, 0x6b,
-	0x65, 0x74, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x5f, 0x0a, 0x12, 0x53, 0x65,
-	0x6e, 0x64, 0x4b, 0x61, 0x6d, 0x69, 0x6b, 0x61, 0x7a, 0x65, 0x50, 0x61, 0x63, 0x6b, 0x65, 0x74,
-	0x12, 0x23, 0x2e, 0x70, 0x72, 0x6f, 0x78, 0x79, 0x2e, 0x76, 0x31, 0x2e, 0x53, 0x65, 0x6e, 0x64,
-	0x4b, 0x61, 0x6d, 0x69, 0x6b, 0x61, 0x7a, 0x65, 0x50, 0x61, 0x63, 0x6b, 0x65, 0x74, 0x52, 0x65,
-	0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x24, 0x2e, 0x70, 0x72, 0x6f, 0x78, 0x79, 0x2e, 0x76, 0x31,
-	0x2e, 0x53, 0x65, 0x6e, 0x64, 0x4b, 0x61, 0x6d, 0x69, 0x6b, 0x61, 0x7a, 0x65, 0x50, 0x61, 0x63,
-	0x6b, 0x65, 0x74, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x42, 0x9f, 0x01, 0x0a, 0x0c,
-	0x63, 0x6f, 0x6d, 0x2e, 0x70, 0x72, 0x6f, 0x78, 0x79, 0x2e, 0x76, 0x31, 0x42, 0x0a, 0x50, 0x72,
-	0x6f, 0x78, 0x79, 0x50, 0x72, 0x6f, 0x74, 0x6f, 0x50, 0x01, 0x5a, 0x42, 0x67, 0x69, 0x74, 0x68,
-	0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x6b, 0x6f, 0x75, 0x73, 0x74, 0x65, 0x63, 0x68, 0x2f,
-	0x6d, 0x61, 0x73, 0x74, 0x65, 0x72, 0x6d, 0x69, 0x6e, 0x64, 0x2f, 0x67, 0x65, 0x6e, 0x2f, 0x70,
-	0x72, 0x6f, 0x74, 0x6f, 0x2f, 0x67, 0x6f, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2f, 0x70, 0x72,
-	0x6f, 0x78, 0x79, 0x2f, 0x76, 0x31, 0x3b, 0x70, 0x72, 0x6f, 0x78, 0x79, 0x76, 0x31, 0xa2, 0x02,
-	0x03, 0x50, 0x58, 0x58, 0xaa, 0x02, 0x08, 0x50, 0x72, 0x6f, 0x78, 0x79, 0x2e, 0x56, 0x31, 0xca,
-	0x02, 0x08, 0x50, 0x72, 0x6f, 0x78, 0x79, 0x5c, 0x56, 0x31, 0xe2, 0x02, 0x14, 0x50, 0x72, 0x6f,
-	0x78, 0x79, 0x5c, 0x56, 0x31, 0x5c, 0x47, 0x50, 0x42, 0x4d, 0x65, 0x74, 0x61, 0x64, 0x61, 0x74,
-	0x61, 0xea, 0x02, 0x09, 0x50, 0x72, 0x6f, 0x78, 0x79, 0x3a, 0x3a, 0x56, 0x31, 0x62, 0x06, 0x70,
-	0x72, 0x6f, 0x74, 0x6f, 0x33,
+	0x6e, 0x73, 0x65, 0x22, 0x45, 0x0a, 0x0d, 0x52, 0x65, 0x64, 0x5a, 0x6f, 0x6e, 0x65, 0x50, 0x61,
+	0x63, 0x6b, 0x65, 0x74, 0x12, 0x10, 0x0a, 0x03, 0x6c, 0x61, 0x74, 0x18, 0x01, 0x20, 0x01, 0x28,
+	0x02, 0x52, 0x03, 0x6c, 0x61, 0x74, 0x12, 0x10, 0x0a, 0x03, 0x6c, 0x6f, 0x6e, 0x18, 0x02, 0x20,
+	0x01, 0x28, 0x02, 0x52, 0x03, 0x6c, 0x6f, 0x6e, 0x12, 0x10, 0x0a, 0x03, 0x72, 0x61, 0x64, 0x18,
+	0x03, 0x20, 0x01, 0x28, 0x02, 0x52, 0x03, 0x72, 0x61, 0x64, 0x22, 0x19, 0x0a, 0x17, 0x47, 0x65,
+	0x74, 0x52, 0x65, 0x64, 0x5a, 0x6f, 0x6e, 0x65, 0x50, 0x61, 0x63, 0x6b, 0x65, 0x74, 0x52, 0x65,
+	0x71, 0x75, 0x65, 0x73, 0x74, 0x22, 0x4f, 0x0a, 0x18, 0x47, 0x65, 0x74, 0x52, 0x65, 0x64, 0x5a,
+	0x6f, 0x6e, 0x65, 0x50, 0x61, 0x63, 0x6b, 0x65, 0x74, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73,
+	0x65, 0x12, 0x33, 0x0a, 0x08, 0x72, 0x65, 0x64, 0x5a, 0x6f, 0x6e, 0x65, 0x73, 0x18, 0x01, 0x20,
+	0x03, 0x28, 0x0b, 0x32, 0x17, 0x2e, 0x70, 0x72, 0x6f, 0x78, 0x79, 0x2e, 0x76, 0x31, 0x2e, 0x52,
+	0x65, 0x64, 0x5a, 0x6f, 0x6e, 0x65, 0x50, 0x61, 0x63, 0x6b, 0x65, 0x74, 0x52, 0x08, 0x72, 0x65,
+	0x64, 0x5a, 0x6f, 0x6e, 0x65, 0x73, 0x32, 0xd0, 0x03, 0x0a, 0x0c, 0x50, 0x72, 0x6f, 0x78, 0x79,
+	0x53, 0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x12, 0x4f, 0x0a, 0x0c, 0x47, 0x65, 0x74, 0x54, 0x65,
+	0x6c, 0x65, 0x6d, 0x65, 0x74, 0x72, 0x79, 0x12, 0x1d, 0x2e, 0x70, 0x72, 0x6f, 0x78, 0x79, 0x2e,
+	0x76, 0x31, 0x2e, 0x47, 0x65, 0x74, 0x54, 0x65, 0x6c, 0x65, 0x6d, 0x65, 0x74, 0x72, 0x79, 0x52,
+	0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1e, 0x2e, 0x70, 0x72, 0x6f, 0x78, 0x79, 0x2e, 0x76,
+	0x31, 0x2e, 0x47, 0x65, 0x74, 0x54, 0x65, 0x6c, 0x65, 0x6d, 0x65, 0x74, 0x72, 0x79, 0x52, 0x65,
+	0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x30, 0x01, 0x12, 0x5e, 0x0a, 0x11, 0x47, 0x65, 0x74, 0x45,
+	0x6e, 0x65, 0x6d, 0x79, 0x54, 0x65, 0x6c, 0x65, 0x6d, 0x65, 0x74, 0x72, 0x79, 0x12, 0x22, 0x2e,
+	0x70, 0x72, 0x6f, 0x78, 0x79, 0x2e, 0x76, 0x31, 0x2e, 0x47, 0x65, 0x74, 0x45, 0x6e, 0x65, 0x6d,
+	0x79, 0x54, 0x65, 0x6c, 0x65, 0x6d, 0x65, 0x74, 0x72, 0x79, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73,
+	0x74, 0x1a, 0x23, 0x2e, 0x70, 0x72, 0x6f, 0x78, 0x79, 0x2e, 0x76, 0x31, 0x2e, 0x47, 0x65, 0x74,
+	0x45, 0x6e, 0x65, 0x6d, 0x79, 0x54, 0x65, 0x6c, 0x65, 0x6d, 0x65, 0x74, 0x72, 0x79, 0x52, 0x65,
+	0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x30, 0x01, 0x12, 0x53, 0x0a, 0x0e, 0x53, 0x65, 0x6e, 0x64,
+	0x4c, 0x6f, 0x63, 0x6b, 0x50, 0x61, 0x63, 0x6b, 0x65, 0x74, 0x12, 0x1f, 0x2e, 0x70, 0x72, 0x6f,
+	0x78, 0x79, 0x2e, 0x76, 0x31, 0x2e, 0x53, 0x65, 0x6e, 0x64, 0x4c, 0x6f, 0x63, 0x6b, 0x50, 0x61,
+	0x63, 0x6b, 0x65, 0x74, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x20, 0x2e, 0x70, 0x72,
+	0x6f, 0x78, 0x79, 0x2e, 0x76, 0x31, 0x2e, 0x53, 0x65, 0x6e, 0x64, 0x4c, 0x6f, 0x63, 0x6b, 0x50,
+	0x61, 0x63, 0x6b, 0x65, 0x74, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x5f, 0x0a,
+	0x12, 0x53, 0x65, 0x6e, 0x64, 0x4b, 0x61, 0x6d, 0x69, 0x6b, 0x61, 0x7a, 0x65, 0x50, 0x61, 0x63,
+	0x6b, 0x65, 0x74, 0x12, 0x23, 0x2e, 0x70, 0x72, 0x6f, 0x78, 0x79, 0x2e, 0x76, 0x31, 0x2e, 0x53,
+	0x65, 0x6e, 0x64, 0x4b, 0x61, 0x6d, 0x69, 0x6b, 0x61, 0x7a, 0x65, 0x50, 0x61, 0x63, 0x6b, 0x65,
+	0x74, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x24, 0x2e, 0x70, 0x72, 0x6f, 0x78, 0x79,
+	0x2e, 0x76, 0x31, 0x2e, 0x53, 0x65, 0x6e, 0x64, 0x4b, 0x61, 0x6d, 0x69, 0x6b, 0x61, 0x7a, 0x65,
+	0x50, 0x61, 0x63, 0x6b, 0x65, 0x74, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x59,
+	0x0a, 0x10, 0x47, 0x65, 0x74, 0x52, 0x65, 0x64, 0x5a, 0x6f, 0x6e, 0x65, 0x50, 0x61, 0x63, 0x6b,
+	0x65, 0x74, 0x12, 0x21, 0x2e, 0x70, 0x72, 0x6f, 0x78, 0x79, 0x2e, 0x76, 0x31, 0x2e, 0x47, 0x65,
+	0x74, 0x52, 0x65, 0x64, 0x5a, 0x6f, 0x6e, 0x65, 0x50, 0x61, 0x63, 0x6b, 0x65, 0x74, 0x52, 0x65,
+	0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x22, 0x2e, 0x70, 0x72, 0x6f, 0x78, 0x79, 0x2e, 0x76, 0x31,
+	0x2e, 0x47, 0x65, 0x74, 0x52, 0x65, 0x64, 0x5a, 0x6f, 0x6e, 0x65, 0x50, 0x61, 0x63, 0x6b, 0x65,
+	0x74, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x42, 0x9f, 0x01, 0x0a, 0x0c, 0x63, 0x6f,
+	0x6d, 0x2e, 0x70, 0x72, 0x6f, 0x78, 0x79, 0x2e, 0x76, 0x31, 0x42, 0x0a, 0x50, 0x72, 0x6f, 0x78,
+	0x79, 0x50, 0x72, 0x6f, 0x74, 0x6f, 0x50, 0x01, 0x5a, 0x42, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62,
+	0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x6b, 0x6f, 0x75, 0x73, 0x74, 0x65, 0x63, 0x68, 0x2f, 0x6d, 0x61,
+	0x73, 0x74, 0x65, 0x72, 0x6d, 0x69, 0x6e, 0x64, 0x2f, 0x67, 0x65, 0x6e, 0x2f, 0x70, 0x72, 0x6f,
+	0x74, 0x6f, 0x2f, 0x67, 0x6f, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2f, 0x70, 0x72, 0x6f, 0x78,
+	0x79, 0x2f, 0x76, 0x31, 0x3b, 0x70, 0x72, 0x6f, 0x78, 0x79, 0x76, 0x31, 0xa2, 0x02, 0x03, 0x50,
+	0x58, 0x58, 0xaa, 0x02, 0x08, 0x50, 0x72, 0x6f, 0x78, 0x79, 0x2e, 0x56, 0x31, 0xca, 0x02, 0x08,
+	0x50, 0x72, 0x6f, 0x78, 0x79, 0x5c, 0x56, 0x31, 0xe2, 0x02, 0x14, 0x50, 0x72, 0x6f, 0x78, 0x79,
+	0x5c, 0x56, 0x31, 0x5c, 0x47, 0x50, 0x42, 0x4d, 0x65, 0x74, 0x61, 0x64, 0x61, 0x74, 0x61, 0xea,
+	0x02, 0x09, 0x50, 0x72, 0x6f, 0x78, 0x79, 0x3a, 0x3a, 0x56, 0x31, 0x62, 0x06, 0x70, 0x72, 0x6f,
+	0x74, 0x6f, 0x33,
 }
 
 var (
@@ -1115,8 +1280,8 @@ func file_proto_proxy_v1_proxy_proto_rawDescGZIP() []byte {
 	return file_proto_proxy_v1_proxy_proto_rawDescData
 }
 
-var file_proto_proxy_v1_proxy_proto_msgTypes = make([]protoimpl.MessageInfo, 13)
-var file_proto_proxy_v1_proxy_proto_goTypes = []interface{}{
+var file_proto_proxy_v1_proxy_proto_msgTypes = make([]protoimpl.MessageInfo, 16)
+var file_proto_proxy_v1_proxy_proto_goTypes = []any{
 	(*GetTelemetryRequest)(nil),          // 0: proxy.v1.GetTelemetryRequest
 	(*GetTelemetryResponse)(nil),         // 1: proxy.v1.GetTelemetryResponse
 	(*Target)(nil),                       // 2: proxy.v1.Target
@@ -1130,6 +1295,9 @@ var file_proto_proxy_v1_proxy_proto_goTypes = []interface{}{
 	(*SendLockPacketResponse)(nil),       // 10: proxy.v1.SendLockPacketResponse
 	(*SendKamikazePacketRequest)(nil),    // 11: proxy.v1.SendKamikazePacketRequest
 	(*SendKamikazePacketResponse)(nil),   // 12: proxy.v1.SendKamikazePacketResponse
+	(*RedZonePacket)(nil),                // 13: proxy.v1.RedZonePacket
+	(*GetRedZonePacketRequest)(nil),      // 14: proxy.v1.GetRedZonePacketRequest
+	(*GetRedZonePacketResponse)(nil),     // 15: proxy.v1.GetRedZonePacketResponse
 }
 var file_proto_proxy_v1_proxy_proto_depIdxs = []int32{
 	2,  // 0: proxy.v1.GetDetailedTelemetryResponse.target:type_name -> proxy.v1.Target
@@ -1139,19 +1307,22 @@ var file_proto_proxy_v1_proxy_proto_depIdxs = []int32{
 	6,  // 4: proxy.v1.SendLockPacketRequest.lock_end_time:type_name -> proxy.v1.Time
 	6,  // 5: proxy.v1.SendKamikazePacketRequest.kamikaze_start_time:type_name -> proxy.v1.Time
 	6,  // 6: proxy.v1.SendKamikazePacketRequest.kamikaze_end_time:type_name -> proxy.v1.Time
-	0,  // 7: proxy.v1.ProxyService.GetTelemetry:input_type -> proxy.v1.GetTelemetryRequest
-	7,  // 8: proxy.v1.ProxyService.GetEnemyTelemetry:input_type -> proxy.v1.GetEnemyTelemetryRequest
-	9,  // 9: proxy.v1.ProxyService.SendLockPacket:input_type -> proxy.v1.SendLockPacketRequest
-	11, // 10: proxy.v1.ProxyService.SendKamikazePacket:input_type -> proxy.v1.SendKamikazePacketRequest
-	1,  // 11: proxy.v1.ProxyService.GetTelemetry:output_type -> proxy.v1.GetTelemetryResponse
-	8,  // 12: proxy.v1.ProxyService.GetEnemyTelemetry:output_type -> proxy.v1.GetEnemyTelemetryResponse
-	10, // 13: proxy.v1.ProxyService.SendLockPacket:output_type -> proxy.v1.SendLockPacketResponse
-	12, // 14: proxy.v1.ProxyService.SendKamikazePacket:output_type -> proxy.v1.SendKamikazePacketResponse
-	11, // [11:15] is the sub-list for method output_type
-	7,  // [7:11] is the sub-list for method input_type
-	7,  // [7:7] is the sub-list for extension type_name
-	7,  // [7:7] is the sub-list for extension extendee
-	0,  // [0:7] is the sub-list for field type_name
+	13, // 7: proxy.v1.GetRedZonePacketResponse.redZones:type_name -> proxy.v1.RedZonePacket
+	0,  // 8: proxy.v1.ProxyService.GetTelemetry:input_type -> proxy.v1.GetTelemetryRequest
+	7,  // 9: proxy.v1.ProxyService.GetEnemyTelemetry:input_type -> proxy.v1.GetEnemyTelemetryRequest
+	9,  // 10: proxy.v1.ProxyService.SendLockPacket:input_type -> proxy.v1.SendLockPacketRequest
+	11, // 11: proxy.v1.ProxyService.SendKamikazePacket:input_type -> proxy.v1.SendKamikazePacketRequest
+	14, // 12: proxy.v1.ProxyService.GetRedZonePacket:input_type -> proxy.v1.GetRedZonePacketRequest
+	1,  // 13: proxy.v1.ProxyService.GetTelemetry:output_type -> proxy.v1.GetTelemetryResponse
+	8,  // 14: proxy.v1.ProxyService.GetEnemyTelemetry:output_type -> proxy.v1.GetEnemyTelemetryResponse
+	10, // 15: proxy.v1.ProxyService.SendLockPacket:output_type -> proxy.v1.SendLockPacketResponse
+	12, // 16: proxy.v1.ProxyService.SendKamikazePacket:output_type -> proxy.v1.SendKamikazePacketResponse
+	15, // 17: proxy.v1.ProxyService.GetRedZonePacket:output_type -> proxy.v1.GetRedZonePacketResponse
+	13, // [13:18] is the sub-list for method output_type
+	8,  // [8:13] is the sub-list for method input_type
+	8,  // [8:8] is the sub-list for extension type_name
+	8,  // [8:8] is the sub-list for extension extendee
+	0,  // [0:8] is the sub-list for field type_name
 }
 
 func init() { file_proto_proxy_v1_proxy_proto_init() }
@@ -1160,7 +1331,7 @@ func file_proto_proxy_v1_proxy_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_proto_proxy_v1_proxy_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_proxy_v1_proxy_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*GetTelemetryRequest); i {
 			case 0:
 				return &v.state
@@ -1172,7 +1343,7 @@ func file_proto_proxy_v1_proxy_proto_init() {
 				return nil
 			}
 		}
-		file_proto_proxy_v1_proxy_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_proxy_v1_proxy_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*GetTelemetryResponse); i {
 			case 0:
 				return &v.state
@@ -1184,7 +1355,7 @@ func file_proto_proxy_v1_proxy_proto_init() {
 				return nil
 			}
 		}
-		file_proto_proxy_v1_proxy_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_proxy_v1_proxy_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*Target); i {
 			case 0:
 				return &v.state
@@ -1196,7 +1367,7 @@ func file_proto_proxy_v1_proxy_proto_init() {
 				return nil
 			}
 		}
-		file_proto_proxy_v1_proxy_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_proxy_v1_proxy_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*GetDetailedTelemetryRequest); i {
 			case 0:
 				return &v.state
@@ -1208,7 +1379,7 @@ func file_proto_proxy_v1_proxy_proto_init() {
 				return nil
 			}
 		}
-		file_proto_proxy_v1_proxy_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_proxy_v1_proxy_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*GetDetailedTelemetryResponse); i {
 			case 0:
 				return &v.state
@@ -1220,7 +1391,7 @@ func file_proto_proxy_v1_proxy_proto_init() {
 				return nil
 			}
 		}
-		file_proto_proxy_v1_proxy_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_proxy_v1_proxy_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*EnemyPlane); i {
 			case 0:
 				return &v.state
@@ -1232,7 +1403,7 @@ func file_proto_proxy_v1_proxy_proto_init() {
 				return nil
 			}
 		}
-		file_proto_proxy_v1_proxy_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_proxy_v1_proxy_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*Time); i {
 			case 0:
 				return &v.state
@@ -1244,7 +1415,7 @@ func file_proto_proxy_v1_proxy_proto_init() {
 				return nil
 			}
 		}
-		file_proto_proxy_v1_proxy_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_proxy_v1_proxy_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*GetEnemyTelemetryRequest); i {
 			case 0:
 				return &v.state
@@ -1256,7 +1427,7 @@ func file_proto_proxy_v1_proxy_proto_init() {
 				return nil
 			}
 		}
-		file_proto_proxy_v1_proxy_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_proxy_v1_proxy_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*GetEnemyTelemetryResponse); i {
 			case 0:
 				return &v.state
@@ -1268,7 +1439,7 @@ func file_proto_proxy_v1_proxy_proto_init() {
 				return nil
 			}
 		}
-		file_proto_proxy_v1_proxy_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_proxy_v1_proxy_proto_msgTypes[9].Exporter = func(v any, i int) any {
 			switch v := v.(*SendLockPacketRequest); i {
 			case 0:
 				return &v.state
@@ -1280,7 +1451,7 @@ func file_proto_proxy_v1_proxy_proto_init() {
 				return nil
 			}
 		}
-		file_proto_proxy_v1_proxy_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_proxy_v1_proxy_proto_msgTypes[10].Exporter = func(v any, i int) any {
 			switch v := v.(*SendLockPacketResponse); i {
 			case 0:
 				return &v.state
@@ -1292,7 +1463,7 @@ func file_proto_proxy_v1_proxy_proto_init() {
 				return nil
 			}
 		}
-		file_proto_proxy_v1_proxy_proto_msgTypes[11].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_proxy_v1_proxy_proto_msgTypes[11].Exporter = func(v any, i int) any {
 			switch v := v.(*SendKamikazePacketRequest); i {
 			case 0:
 				return &v.state
@@ -1304,7 +1475,7 @@ func file_proto_proxy_v1_proxy_proto_init() {
 				return nil
 			}
 		}
-		file_proto_proxy_v1_proxy_proto_msgTypes[12].Exporter = func(v interface{}, i int) interface{} {
+		file_proto_proxy_v1_proxy_proto_msgTypes[12].Exporter = func(v any, i int) any {
 			switch v := v.(*SendKamikazePacketResponse); i {
 			case 0:
 				return &v.state
@@ -1316,15 +1487,51 @@ func file_proto_proxy_v1_proxy_proto_init() {
 				return nil
 			}
 		}
+		file_proto_proxy_v1_proxy_proto_msgTypes[13].Exporter = func(v any, i int) any {
+			switch v := v.(*RedZonePacket); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_proto_proxy_v1_proxy_proto_msgTypes[14].Exporter = func(v any, i int) any {
+			switch v := v.(*GetRedZonePacketRequest); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
+		file_proto_proxy_v1_proxy_proto_msgTypes[15].Exporter = func(v any, i int) any {
+			switch v := v.(*GetRedZonePacketResponse); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
 	}
-	file_proto_proxy_v1_proxy_proto_msgTypes[4].OneofWrappers = []interface{}{}
+	file_proto_proxy_v1_proxy_proto_msgTypes[4].OneofWrappers = []any{}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
 			RawDescriptor: file_proto_proxy_v1_proxy_proto_rawDesc,
 			NumEnums:      0,
-			NumMessages:   13,
+			NumMessages:   16,
 			NumExtensions: 0,
 			NumServices:   1,
 		},
