@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
+// - protoc-gen-go-grpc v1.5.1
 // - protoc             (unknown)
 // source: proto/approach/v1/approach.proto
 
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	ApproachService_GetPlanes_FullMethodName     = "/approach.v1.ApproachService/GetPlanes"
@@ -26,11 +26,14 @@ const (
 // ApproachServiceClient is the client API for ApproachService service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// ApproachService is responsible for calculating and giving commands to
+// KOUSTECH's UAV to approach a target enemy UAV
 type ApproachServiceClient interface {
 	// A server-streaming endpoint
 	// Accepts a GetPlanesRequest and returns a stream of airplanes and their
 	// properties, and the selected plane id if it exists
-	GetPlanes(ctx context.Context, in *GetPlanesRequest, opts ...grpc.CallOption) (ApproachService_GetPlanesClient, error)
+	GetPlanes(ctx context.Context, in *GetPlanesRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[GetPlanesResponse], error)
 	// ApproachPlane sends a command to approach the plane with specified ID
 	ApproachPlane(ctx context.Context, in *ApproachPlaneRequest, opts ...grpc.CallOption) (*ApproachPlaneResponse, error)
 }
@@ -43,12 +46,13 @@ func NewApproachServiceClient(cc grpc.ClientConnInterface) ApproachServiceClient
 	return &approachServiceClient{cc}
 }
 
-func (c *approachServiceClient) GetPlanes(ctx context.Context, in *GetPlanesRequest, opts ...grpc.CallOption) (ApproachService_GetPlanesClient, error) {
-	stream, err := c.cc.NewStream(ctx, &ApproachService_ServiceDesc.Streams[0], ApproachService_GetPlanes_FullMethodName, opts...)
+func (c *approachServiceClient) GetPlanes(ctx context.Context, in *GetPlanesRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[GetPlanesResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &ApproachService_ServiceDesc.Streams[0], ApproachService_GetPlanes_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &approachServiceGetPlanesClient{stream}
+	x := &grpc.GenericClientStream[GetPlanesRequest, GetPlanesResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -58,26 +62,13 @@ func (c *approachServiceClient) GetPlanes(ctx context.Context, in *GetPlanesRequ
 	return x, nil
 }
 
-type ApproachService_GetPlanesClient interface {
-	Recv() (*GetPlanesResponse, error)
-	grpc.ClientStream
-}
-
-type approachServiceGetPlanesClient struct {
-	grpc.ClientStream
-}
-
-func (x *approachServiceGetPlanesClient) Recv() (*GetPlanesResponse, error) {
-	m := new(GetPlanesResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type ApproachService_GetPlanesClient = grpc.ServerStreamingClient[GetPlanesResponse]
 
 func (c *approachServiceClient) ApproachPlane(ctx context.Context, in *ApproachPlaneRequest, opts ...grpc.CallOption) (*ApproachPlaneResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ApproachPlaneResponse)
-	err := c.cc.Invoke(ctx, ApproachService_ApproachPlane_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ApproachService_ApproachPlane_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -86,26 +77,33 @@ func (c *approachServiceClient) ApproachPlane(ctx context.Context, in *ApproachP
 
 // ApproachServiceServer is the server API for ApproachService service.
 // All implementations should embed UnimplementedApproachServiceServer
-// for forward compatibility
+// for forward compatibility.
+//
+// ApproachService is responsible for calculating and giving commands to
+// KOUSTECH's UAV to approach a target enemy UAV
 type ApproachServiceServer interface {
 	// A server-streaming endpoint
 	// Accepts a GetPlanesRequest and returns a stream of airplanes and their
 	// properties, and the selected plane id if it exists
-	GetPlanes(*GetPlanesRequest, ApproachService_GetPlanesServer) error
+	GetPlanes(*GetPlanesRequest, grpc.ServerStreamingServer[GetPlanesResponse]) error
 	// ApproachPlane sends a command to approach the plane with specified ID
 	ApproachPlane(context.Context, *ApproachPlaneRequest) (*ApproachPlaneResponse, error)
 }
 
-// UnimplementedApproachServiceServer should be embedded to have forward compatible implementations.
-type UnimplementedApproachServiceServer struct {
-}
+// UnimplementedApproachServiceServer should be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedApproachServiceServer struct{}
 
-func (UnimplementedApproachServiceServer) GetPlanes(*GetPlanesRequest, ApproachService_GetPlanesServer) error {
+func (UnimplementedApproachServiceServer) GetPlanes(*GetPlanesRequest, grpc.ServerStreamingServer[GetPlanesResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method GetPlanes not implemented")
 }
 func (UnimplementedApproachServiceServer) ApproachPlane(context.Context, *ApproachPlaneRequest) (*ApproachPlaneResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ApproachPlane not implemented")
 }
+func (UnimplementedApproachServiceServer) testEmbeddedByValue() {}
 
 // UnsafeApproachServiceServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to ApproachServiceServer will
@@ -115,6 +113,13 @@ type UnsafeApproachServiceServer interface {
 }
 
 func RegisterApproachServiceServer(s grpc.ServiceRegistrar, srv ApproachServiceServer) {
+	// If the following call pancis, it indicates UnimplementedApproachServiceServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&ApproachService_ServiceDesc, srv)
 }
 
@@ -123,21 +128,11 @@ func _ApproachService_GetPlanes_Handler(srv interface{}, stream grpc.ServerStrea
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(ApproachServiceServer).GetPlanes(m, &approachServiceGetPlanesServer{stream})
+	return srv.(ApproachServiceServer).GetPlanes(m, &grpc.GenericServerStream[GetPlanesRequest, GetPlanesResponse]{ServerStream: stream})
 }
 
-type ApproachService_GetPlanesServer interface {
-	Send(*GetPlanesResponse) error
-	grpc.ServerStream
-}
-
-type approachServiceGetPlanesServer struct {
-	grpc.ServerStream
-}
-
-func (x *approachServiceGetPlanesServer) Send(m *GetPlanesResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type ApproachService_GetPlanesServer = grpc.ServerStreamingServer[GetPlanesResponse]
 
 func _ApproachService_ApproachPlane_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(ApproachPlaneRequest)
