@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
+// - protoc-gen-go-grpc v1.5.1
 // - protoc             (unknown)
 // source: proto/mastermind/v1/mastermind.proto
 
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	MastermindService_UpdateState_FullMethodName          = "/mastermind.v1.MastermindService/UpdateState"
@@ -36,11 +36,11 @@ type MastermindServiceClient interface {
 	// Accepts a stream of UpdateStateRequests during the competition, while
 	// sending UpdateStateResponses whenever a valid StateTransition is submitted
 	// by any client
-	UpdateState(ctx context.Context, opts ...grpc.CallOption) (MastermindService_UpdateStateClient, error)
-	GetTelemetry(ctx context.Context, in *GetTelemetryRequest, opts ...grpc.CallOption) (MastermindService_GetTelemetryClient, error)
+	UpdateState(ctx context.Context, opts ...grpc.CallOption) (grpc.BidiStreamingClient[UpdateStateRequest, UpdateStateResponse], error)
+	GetTelemetry(ctx context.Context, in *GetTelemetryRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[GetTelemetryResponse], error)
 	// Returns a detailed telemetry response. SHOULD be used by services that
 	// interact with the competiiton server
-	GetDetailedTelemetry(ctx context.Context, in *GetDetailedTelemetryRequest, opts ...grpc.CallOption) (MastermindService_GetDetailedTelemetryClient, error)
+	GetDetailedTelemetry(ctx context.Context, in *GetDetailedTelemetryRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[GetDetailedTelemetryResponse], error)
 	// SetPIDLevel sets the PID level for the plane
 	SetPIDLevel(ctx context.Context, in *SetPIDLevelRequest, opts ...grpc.CallOption) (*SetPIDLevelResponse, error)
 	// SetSpeed sets the plane's airspeed in m/s
@@ -48,7 +48,7 @@ type MastermindServiceClient interface {
 	// GotoWaypoint commands the plane to go to specified waypoint
 	GotoWaypoint(ctx context.Context, in *GotoWaypointRequest, opts ...grpc.CallOption) (*GotoWaypointResponse, error)
 	// SetAttitude comannds the plane to change its attitude
-	SetAttitude(ctx context.Context, opts ...grpc.CallOption) (MastermindService_SetAttitudeClient, error)
+	SetAttitude(ctx context.Context, opts ...grpc.CallOption) (grpc.ClientStreamingClient[SetAttitudeRequest, SetAttitudeResponse], error)
 }
 
 type mastermindServiceClient struct {
@@ -59,43 +59,26 @@ func NewMastermindServiceClient(cc grpc.ClientConnInterface) MastermindServiceCl
 	return &mastermindServiceClient{cc}
 }
 
-func (c *mastermindServiceClient) UpdateState(ctx context.Context, opts ...grpc.CallOption) (MastermindService_UpdateStateClient, error) {
-	stream, err := c.cc.NewStream(ctx, &MastermindService_ServiceDesc.Streams[0], MastermindService_UpdateState_FullMethodName, opts...)
+func (c *mastermindServiceClient) UpdateState(ctx context.Context, opts ...grpc.CallOption) (grpc.BidiStreamingClient[UpdateStateRequest, UpdateStateResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &MastermindService_ServiceDesc.Streams[0], MastermindService_UpdateState_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &mastermindServiceUpdateStateClient{stream}
+	x := &grpc.GenericClientStream[UpdateStateRequest, UpdateStateResponse]{ClientStream: stream}
 	return x, nil
 }
 
-type MastermindService_UpdateStateClient interface {
-	Send(*UpdateStateRequest) error
-	Recv() (*UpdateStateResponse, error)
-	grpc.ClientStream
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type MastermindService_UpdateStateClient = grpc.BidiStreamingClient[UpdateStateRequest, UpdateStateResponse]
 
-type mastermindServiceUpdateStateClient struct {
-	grpc.ClientStream
-}
-
-func (x *mastermindServiceUpdateStateClient) Send(m *UpdateStateRequest) error {
-	return x.ClientStream.SendMsg(m)
-}
-
-func (x *mastermindServiceUpdateStateClient) Recv() (*UpdateStateResponse, error) {
-	m := new(UpdateStateResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
-
-func (c *mastermindServiceClient) GetTelemetry(ctx context.Context, in *GetTelemetryRequest, opts ...grpc.CallOption) (MastermindService_GetTelemetryClient, error) {
-	stream, err := c.cc.NewStream(ctx, &MastermindService_ServiceDesc.Streams[1], MastermindService_GetTelemetry_FullMethodName, opts...)
+func (c *mastermindServiceClient) GetTelemetry(ctx context.Context, in *GetTelemetryRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[GetTelemetryResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &MastermindService_ServiceDesc.Streams[1], MastermindService_GetTelemetry_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &mastermindServiceGetTelemetryClient{stream}
+	x := &grpc.GenericClientStream[GetTelemetryRequest, GetTelemetryResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -105,29 +88,16 @@ func (c *mastermindServiceClient) GetTelemetry(ctx context.Context, in *GetTelem
 	return x, nil
 }
 
-type MastermindService_GetTelemetryClient interface {
-	Recv() (*GetTelemetryResponse, error)
-	grpc.ClientStream
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type MastermindService_GetTelemetryClient = grpc.ServerStreamingClient[GetTelemetryResponse]
 
-type mastermindServiceGetTelemetryClient struct {
-	grpc.ClientStream
-}
-
-func (x *mastermindServiceGetTelemetryClient) Recv() (*GetTelemetryResponse, error) {
-	m := new(GetTelemetryResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
-
-func (c *mastermindServiceClient) GetDetailedTelemetry(ctx context.Context, in *GetDetailedTelemetryRequest, opts ...grpc.CallOption) (MastermindService_GetDetailedTelemetryClient, error) {
-	stream, err := c.cc.NewStream(ctx, &MastermindService_ServiceDesc.Streams[2], MastermindService_GetDetailedTelemetry_FullMethodName, opts...)
+func (c *mastermindServiceClient) GetDetailedTelemetry(ctx context.Context, in *GetDetailedTelemetryRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[GetDetailedTelemetryResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &MastermindService_ServiceDesc.Streams[2], MastermindService_GetDetailedTelemetry_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &mastermindServiceGetDetailedTelemetryClient{stream}
+	x := &grpc.GenericClientStream[GetDetailedTelemetryRequest, GetDetailedTelemetryResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -137,26 +107,13 @@ func (c *mastermindServiceClient) GetDetailedTelemetry(ctx context.Context, in *
 	return x, nil
 }
 
-type MastermindService_GetDetailedTelemetryClient interface {
-	Recv() (*GetDetailedTelemetryResponse, error)
-	grpc.ClientStream
-}
-
-type mastermindServiceGetDetailedTelemetryClient struct {
-	grpc.ClientStream
-}
-
-func (x *mastermindServiceGetDetailedTelemetryClient) Recv() (*GetDetailedTelemetryResponse, error) {
-	m := new(GetDetailedTelemetryResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type MastermindService_GetDetailedTelemetryClient = grpc.ServerStreamingClient[GetDetailedTelemetryResponse]
 
 func (c *mastermindServiceClient) SetPIDLevel(ctx context.Context, in *SetPIDLevelRequest, opts ...grpc.CallOption) (*SetPIDLevelResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(SetPIDLevelResponse)
-	err := c.cc.Invoke(ctx, MastermindService_SetPIDLevel_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MastermindService_SetPIDLevel_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -164,8 +121,9 @@ func (c *mastermindServiceClient) SetPIDLevel(ctx context.Context, in *SetPIDLev
 }
 
 func (c *mastermindServiceClient) SetSpeed(ctx context.Context, in *SetSpeedRequest, opts ...grpc.CallOption) (*SetSpeedResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(SetSpeedResponse)
-	err := c.cc.Invoke(ctx, MastermindService_SetSpeed_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MastermindService_SetSpeed_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -173,61 +131,41 @@ func (c *mastermindServiceClient) SetSpeed(ctx context.Context, in *SetSpeedRequ
 }
 
 func (c *mastermindServiceClient) GotoWaypoint(ctx context.Context, in *GotoWaypointRequest, opts ...grpc.CallOption) (*GotoWaypointResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(GotoWaypointResponse)
-	err := c.cc.Invoke(ctx, MastermindService_GotoWaypoint_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MastermindService_GotoWaypoint_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *mastermindServiceClient) SetAttitude(ctx context.Context, opts ...grpc.CallOption) (MastermindService_SetAttitudeClient, error) {
-	stream, err := c.cc.NewStream(ctx, &MastermindService_ServiceDesc.Streams[3], MastermindService_SetAttitude_FullMethodName, opts...)
+func (c *mastermindServiceClient) SetAttitude(ctx context.Context, opts ...grpc.CallOption) (grpc.ClientStreamingClient[SetAttitudeRequest, SetAttitudeResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &MastermindService_ServiceDesc.Streams[3], MastermindService_SetAttitude_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &mastermindServiceSetAttitudeClient{stream}
+	x := &grpc.GenericClientStream[SetAttitudeRequest, SetAttitudeResponse]{ClientStream: stream}
 	return x, nil
 }
 
-type MastermindService_SetAttitudeClient interface {
-	Send(*SetAttitudeRequest) error
-	CloseAndRecv() (*SetAttitudeResponse, error)
-	grpc.ClientStream
-}
-
-type mastermindServiceSetAttitudeClient struct {
-	grpc.ClientStream
-}
-
-func (x *mastermindServiceSetAttitudeClient) Send(m *SetAttitudeRequest) error {
-	return x.ClientStream.SendMsg(m)
-}
-
-func (x *mastermindServiceSetAttitudeClient) CloseAndRecv() (*SetAttitudeResponse, error) {
-	if err := x.ClientStream.CloseSend(); err != nil {
-		return nil, err
-	}
-	m := new(SetAttitudeResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type MastermindService_SetAttitudeClient = grpc.ClientStreamingClient[SetAttitudeRequest, SetAttitudeResponse]
 
 // MastermindServiceServer is the server API for MastermindService service.
 // All implementations should embed UnimplementedMastermindServiceServer
-// for forward compatibility
+// for forward compatibility.
 type MastermindServiceServer interface {
 	// A bidirectional stream
 	// Accepts a stream of UpdateStateRequests during the competition, while
 	// sending UpdateStateResponses whenever a valid StateTransition is submitted
 	// by any client
-	UpdateState(MastermindService_UpdateStateServer) error
-	GetTelemetry(*GetTelemetryRequest, MastermindService_GetTelemetryServer) error
+	UpdateState(grpc.BidiStreamingServer[UpdateStateRequest, UpdateStateResponse]) error
+	GetTelemetry(*GetTelemetryRequest, grpc.ServerStreamingServer[GetTelemetryResponse]) error
 	// Returns a detailed telemetry response. SHOULD be used by services that
 	// interact with the competiiton server
-	GetDetailedTelemetry(*GetDetailedTelemetryRequest, MastermindService_GetDetailedTelemetryServer) error
+	GetDetailedTelemetry(*GetDetailedTelemetryRequest, grpc.ServerStreamingServer[GetDetailedTelemetryResponse]) error
 	// SetPIDLevel sets the PID level for the plane
 	SetPIDLevel(context.Context, *SetPIDLevelRequest) (*SetPIDLevelResponse, error)
 	// SetSpeed sets the plane's airspeed in m/s
@@ -235,20 +173,23 @@ type MastermindServiceServer interface {
 	// GotoWaypoint commands the plane to go to specified waypoint
 	GotoWaypoint(context.Context, *GotoWaypointRequest) (*GotoWaypointResponse, error)
 	// SetAttitude comannds the plane to change its attitude
-	SetAttitude(MastermindService_SetAttitudeServer) error
+	SetAttitude(grpc.ClientStreamingServer[SetAttitudeRequest, SetAttitudeResponse]) error
 }
 
-// UnimplementedMastermindServiceServer should be embedded to have forward compatible implementations.
-type UnimplementedMastermindServiceServer struct {
-}
+// UnimplementedMastermindServiceServer should be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedMastermindServiceServer struct{}
 
-func (UnimplementedMastermindServiceServer) UpdateState(MastermindService_UpdateStateServer) error {
+func (UnimplementedMastermindServiceServer) UpdateState(grpc.BidiStreamingServer[UpdateStateRequest, UpdateStateResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method UpdateState not implemented")
 }
-func (UnimplementedMastermindServiceServer) GetTelemetry(*GetTelemetryRequest, MastermindService_GetTelemetryServer) error {
+func (UnimplementedMastermindServiceServer) GetTelemetry(*GetTelemetryRequest, grpc.ServerStreamingServer[GetTelemetryResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method GetTelemetry not implemented")
 }
-func (UnimplementedMastermindServiceServer) GetDetailedTelemetry(*GetDetailedTelemetryRequest, MastermindService_GetDetailedTelemetryServer) error {
+func (UnimplementedMastermindServiceServer) GetDetailedTelemetry(*GetDetailedTelemetryRequest, grpc.ServerStreamingServer[GetDetailedTelemetryResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method GetDetailedTelemetry not implemented")
 }
 func (UnimplementedMastermindServiceServer) SetPIDLevel(context.Context, *SetPIDLevelRequest) (*SetPIDLevelResponse, error) {
@@ -260,9 +201,10 @@ func (UnimplementedMastermindServiceServer) SetSpeed(context.Context, *SetSpeedR
 func (UnimplementedMastermindServiceServer) GotoWaypoint(context.Context, *GotoWaypointRequest) (*GotoWaypointResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GotoWaypoint not implemented")
 }
-func (UnimplementedMastermindServiceServer) SetAttitude(MastermindService_SetAttitudeServer) error {
+func (UnimplementedMastermindServiceServer) SetAttitude(grpc.ClientStreamingServer[SetAttitudeRequest, SetAttitudeResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method SetAttitude not implemented")
 }
+func (UnimplementedMastermindServiceServer) testEmbeddedByValue() {}
 
 // UnsafeMastermindServiceServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to MastermindServiceServer will
@@ -272,76 +214,44 @@ type UnsafeMastermindServiceServer interface {
 }
 
 func RegisterMastermindServiceServer(s grpc.ServiceRegistrar, srv MastermindServiceServer) {
+	// If the following call pancis, it indicates UnimplementedMastermindServiceServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&MastermindService_ServiceDesc, srv)
 }
 
 func _MastermindService_UpdateState_Handler(srv interface{}, stream grpc.ServerStream) error {
-	return srv.(MastermindServiceServer).UpdateState(&mastermindServiceUpdateStateServer{stream})
+	return srv.(MastermindServiceServer).UpdateState(&grpc.GenericServerStream[UpdateStateRequest, UpdateStateResponse]{ServerStream: stream})
 }
 
-type MastermindService_UpdateStateServer interface {
-	Send(*UpdateStateResponse) error
-	Recv() (*UpdateStateRequest, error)
-	grpc.ServerStream
-}
-
-type mastermindServiceUpdateStateServer struct {
-	grpc.ServerStream
-}
-
-func (x *mastermindServiceUpdateStateServer) Send(m *UpdateStateResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
-
-func (x *mastermindServiceUpdateStateServer) Recv() (*UpdateStateRequest, error) {
-	m := new(UpdateStateRequest)
-	if err := x.ServerStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type MastermindService_UpdateStateServer = grpc.BidiStreamingServer[UpdateStateRequest, UpdateStateResponse]
 
 func _MastermindService_GetTelemetry_Handler(srv interface{}, stream grpc.ServerStream) error {
 	m := new(GetTelemetryRequest)
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(MastermindServiceServer).GetTelemetry(m, &mastermindServiceGetTelemetryServer{stream})
+	return srv.(MastermindServiceServer).GetTelemetry(m, &grpc.GenericServerStream[GetTelemetryRequest, GetTelemetryResponse]{ServerStream: stream})
 }
 
-type MastermindService_GetTelemetryServer interface {
-	Send(*GetTelemetryResponse) error
-	grpc.ServerStream
-}
-
-type mastermindServiceGetTelemetryServer struct {
-	grpc.ServerStream
-}
-
-func (x *mastermindServiceGetTelemetryServer) Send(m *GetTelemetryResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type MastermindService_GetTelemetryServer = grpc.ServerStreamingServer[GetTelemetryResponse]
 
 func _MastermindService_GetDetailedTelemetry_Handler(srv interface{}, stream grpc.ServerStream) error {
 	m := new(GetDetailedTelemetryRequest)
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(MastermindServiceServer).GetDetailedTelemetry(m, &mastermindServiceGetDetailedTelemetryServer{stream})
+	return srv.(MastermindServiceServer).GetDetailedTelemetry(m, &grpc.GenericServerStream[GetDetailedTelemetryRequest, GetDetailedTelemetryResponse]{ServerStream: stream})
 }
 
-type MastermindService_GetDetailedTelemetryServer interface {
-	Send(*GetDetailedTelemetryResponse) error
-	grpc.ServerStream
-}
-
-type mastermindServiceGetDetailedTelemetryServer struct {
-	grpc.ServerStream
-}
-
-func (x *mastermindServiceGetDetailedTelemetryServer) Send(m *GetDetailedTelemetryResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type MastermindService_GetDetailedTelemetryServer = grpc.ServerStreamingServer[GetDetailedTelemetryResponse]
 
 func _MastermindService_SetPIDLevel_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(SetPIDLevelRequest)
@@ -398,30 +308,11 @@ func _MastermindService_GotoWaypoint_Handler(srv interface{}, ctx context.Contex
 }
 
 func _MastermindService_SetAttitude_Handler(srv interface{}, stream grpc.ServerStream) error {
-	return srv.(MastermindServiceServer).SetAttitude(&mastermindServiceSetAttitudeServer{stream})
+	return srv.(MastermindServiceServer).SetAttitude(&grpc.GenericServerStream[SetAttitudeRequest, SetAttitudeResponse]{ServerStream: stream})
 }
 
-type MastermindService_SetAttitudeServer interface {
-	SendAndClose(*SetAttitudeResponse) error
-	Recv() (*SetAttitudeRequest, error)
-	grpc.ServerStream
-}
-
-type mastermindServiceSetAttitudeServer struct {
-	grpc.ServerStream
-}
-
-func (x *mastermindServiceSetAttitudeServer) SendAndClose(m *SetAttitudeResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
-
-func (x *mastermindServiceSetAttitudeServer) Recv() (*SetAttitudeRequest, error) {
-	m := new(SetAttitudeRequest)
-	if err := x.ServerStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type MastermindService_SetAttitudeServer = grpc.ClientStreamingServer[SetAttitudeRequest, SetAttitudeResponse]
 
 // MastermindService_ServiceDesc is the grpc.ServiceDesc for MastermindService service.
 // It's only intended for direct use with grpc.RegisterService,
