@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
+// - protoc-gen-go-grpc v1.5.1
 // - protoc             (unknown)
 // source: proto/proxy/v1/proxy.proto
 
@@ -15,14 +15,15 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	ProxyService_GetTelemetry_FullMethodName       = "/proxy.v1.ProxyService/GetTelemetry"
 	ProxyService_GetEnemyTelemetry_FullMethodName  = "/proxy.v1.ProxyService/GetEnemyTelemetry"
 	ProxyService_SendLockPacket_FullMethodName     = "/proxy.v1.ProxyService/SendLockPacket"
 	ProxyService_SendKamikazePacket_FullMethodName = "/proxy.v1.ProxyService/SendKamikazePacket"
+	ProxyService_GetRedZonePacket_FullMethodName   = "/proxy.v1.ProxyService/GetRedZonePacket"
 )
 
 // ProxyServiceClient is the client API for ProxyService service.
@@ -31,14 +32,15 @@ const (
 type ProxyServiceClient interface {
 	// Returns a normal telemetry stream when activated. Should be used by
 	// services that don't interact with the competition server
-	GetTelemetry(ctx context.Context, in *GetTelemetryRequest, opts ...grpc.CallOption) (ProxyService_GetTelemetryClient, error)
+	GetTelemetry(ctx context.Context, in *GetTelemetryRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[GetTelemetryResponse], error)
 	// GetEnemyTelemetry returns the enemy's telemetry stream when
 	// activated
-	GetEnemyTelemetry(ctx context.Context, in *GetEnemyTelemetryRequest, opts ...grpc.CallOption) (ProxyService_GetEnemyTelemetryClient, error)
+	GetEnemyTelemetry(ctx context.Context, in *GetEnemyTelemetryRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[GetEnemyTelemetryResponse], error)
 	// SendLockPacket sends a successful lock notification to the competition server
 	SendLockPacket(ctx context.Context, in *SendLockPacketRequest, opts ...grpc.CallOption) (*SendLockPacketResponse, error)
 	// SendKamikazePacket sends a successful kamikaze QR code reading to the competition server
 	SendKamikazePacket(ctx context.Context, in *SendKamikazePacketRequest, opts ...grpc.CallOption) (*SendKamikazePacketResponse, error)
+	GetRedZonePacket(ctx context.Context, in *GetRedZonePacketRequest, opts ...grpc.CallOption) (*GetRedZonePacketResponse, error)
 }
 
 type proxyServiceClient struct {
@@ -49,12 +51,13 @@ func NewProxyServiceClient(cc grpc.ClientConnInterface) ProxyServiceClient {
 	return &proxyServiceClient{cc}
 }
 
-func (c *proxyServiceClient) GetTelemetry(ctx context.Context, in *GetTelemetryRequest, opts ...grpc.CallOption) (ProxyService_GetTelemetryClient, error) {
-	stream, err := c.cc.NewStream(ctx, &ProxyService_ServiceDesc.Streams[0], ProxyService_GetTelemetry_FullMethodName, opts...)
+func (c *proxyServiceClient) GetTelemetry(ctx context.Context, in *GetTelemetryRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[GetTelemetryResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &ProxyService_ServiceDesc.Streams[0], ProxyService_GetTelemetry_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &proxyServiceGetTelemetryClient{stream}
+	x := &grpc.GenericClientStream[GetTelemetryRequest, GetTelemetryResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -64,29 +67,16 @@ func (c *proxyServiceClient) GetTelemetry(ctx context.Context, in *GetTelemetryR
 	return x, nil
 }
 
-type ProxyService_GetTelemetryClient interface {
-	Recv() (*GetTelemetryResponse, error)
-	grpc.ClientStream
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type ProxyService_GetTelemetryClient = grpc.ServerStreamingClient[GetTelemetryResponse]
 
-type proxyServiceGetTelemetryClient struct {
-	grpc.ClientStream
-}
-
-func (x *proxyServiceGetTelemetryClient) Recv() (*GetTelemetryResponse, error) {
-	m := new(GetTelemetryResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
-
-func (c *proxyServiceClient) GetEnemyTelemetry(ctx context.Context, in *GetEnemyTelemetryRequest, opts ...grpc.CallOption) (ProxyService_GetEnemyTelemetryClient, error) {
-	stream, err := c.cc.NewStream(ctx, &ProxyService_ServiceDesc.Streams[1], ProxyService_GetEnemyTelemetry_FullMethodName, opts...)
+func (c *proxyServiceClient) GetEnemyTelemetry(ctx context.Context, in *GetEnemyTelemetryRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[GetEnemyTelemetryResponse], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &ProxyService_ServiceDesc.Streams[1], ProxyService_GetEnemyTelemetry_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &proxyServiceGetEnemyTelemetryClient{stream}
+	x := &grpc.GenericClientStream[GetEnemyTelemetryRequest, GetEnemyTelemetryResponse]{ClientStream: stream}
 	if err := x.ClientStream.SendMsg(in); err != nil {
 		return nil, err
 	}
@@ -96,26 +86,13 @@ func (c *proxyServiceClient) GetEnemyTelemetry(ctx context.Context, in *GetEnemy
 	return x, nil
 }
 
-type ProxyService_GetEnemyTelemetryClient interface {
-	Recv() (*GetEnemyTelemetryResponse, error)
-	grpc.ClientStream
-}
-
-type proxyServiceGetEnemyTelemetryClient struct {
-	grpc.ClientStream
-}
-
-func (x *proxyServiceGetEnemyTelemetryClient) Recv() (*GetEnemyTelemetryResponse, error) {
-	m := new(GetEnemyTelemetryResponse)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type ProxyService_GetEnemyTelemetryClient = grpc.ServerStreamingClient[GetEnemyTelemetryResponse]
 
 func (c *proxyServiceClient) SendLockPacket(ctx context.Context, in *SendLockPacketRequest, opts ...grpc.CallOption) (*SendLockPacketResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(SendLockPacketResponse)
-	err := c.cc.Invoke(ctx, ProxyService_SendLockPacket_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ProxyService_SendLockPacket_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -123,8 +100,19 @@ func (c *proxyServiceClient) SendLockPacket(ctx context.Context, in *SendLockPac
 }
 
 func (c *proxyServiceClient) SendKamikazePacket(ctx context.Context, in *SendKamikazePacketRequest, opts ...grpc.CallOption) (*SendKamikazePacketResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(SendKamikazePacketResponse)
-	err := c.cc.Invoke(ctx, ProxyService_SendKamikazePacket_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, ProxyService_SendKamikazePacket_FullMethodName, in, out, cOpts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *proxyServiceClient) GetRedZonePacket(ctx context.Context, in *GetRedZonePacketRequest, opts ...grpc.CallOption) (*GetRedZonePacketResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(GetRedZonePacketResponse)
+	err := c.cc.Invoke(ctx, ProxyService_GetRedZonePacket_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -133,28 +121,32 @@ func (c *proxyServiceClient) SendKamikazePacket(ctx context.Context, in *SendKam
 
 // ProxyServiceServer is the server API for ProxyService service.
 // All implementations should embed UnimplementedProxyServiceServer
-// for forward compatibility
+// for forward compatibility.
 type ProxyServiceServer interface {
 	// Returns a normal telemetry stream when activated. Should be used by
 	// services that don't interact with the competition server
-	GetTelemetry(*GetTelemetryRequest, ProxyService_GetTelemetryServer) error
+	GetTelemetry(*GetTelemetryRequest, grpc.ServerStreamingServer[GetTelemetryResponse]) error
 	// GetEnemyTelemetry returns the enemy's telemetry stream when
 	// activated
-	GetEnemyTelemetry(*GetEnemyTelemetryRequest, ProxyService_GetEnemyTelemetryServer) error
+	GetEnemyTelemetry(*GetEnemyTelemetryRequest, grpc.ServerStreamingServer[GetEnemyTelemetryResponse]) error
 	// SendLockPacket sends a successful lock notification to the competition server
 	SendLockPacket(context.Context, *SendLockPacketRequest) (*SendLockPacketResponse, error)
 	// SendKamikazePacket sends a successful kamikaze QR code reading to the competition server
 	SendKamikazePacket(context.Context, *SendKamikazePacketRequest) (*SendKamikazePacketResponse, error)
+	GetRedZonePacket(context.Context, *GetRedZonePacketRequest) (*GetRedZonePacketResponse, error)
 }
 
-// UnimplementedProxyServiceServer should be embedded to have forward compatible implementations.
-type UnimplementedProxyServiceServer struct {
-}
+// UnimplementedProxyServiceServer should be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedProxyServiceServer struct{}
 
-func (UnimplementedProxyServiceServer) GetTelemetry(*GetTelemetryRequest, ProxyService_GetTelemetryServer) error {
+func (UnimplementedProxyServiceServer) GetTelemetry(*GetTelemetryRequest, grpc.ServerStreamingServer[GetTelemetryResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method GetTelemetry not implemented")
 }
-func (UnimplementedProxyServiceServer) GetEnemyTelemetry(*GetEnemyTelemetryRequest, ProxyService_GetEnemyTelemetryServer) error {
+func (UnimplementedProxyServiceServer) GetEnemyTelemetry(*GetEnemyTelemetryRequest, grpc.ServerStreamingServer[GetEnemyTelemetryResponse]) error {
 	return status.Errorf(codes.Unimplemented, "method GetEnemyTelemetry not implemented")
 }
 func (UnimplementedProxyServiceServer) SendLockPacket(context.Context, *SendLockPacketRequest) (*SendLockPacketResponse, error) {
@@ -163,6 +155,10 @@ func (UnimplementedProxyServiceServer) SendLockPacket(context.Context, *SendLock
 func (UnimplementedProxyServiceServer) SendKamikazePacket(context.Context, *SendKamikazePacketRequest) (*SendKamikazePacketResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method SendKamikazePacket not implemented")
 }
+func (UnimplementedProxyServiceServer) GetRedZonePacket(context.Context, *GetRedZonePacketRequest) (*GetRedZonePacketResponse, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method GetRedZonePacket not implemented")
+}
+func (UnimplementedProxyServiceServer) testEmbeddedByValue() {}
 
 // UnsafeProxyServiceServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to ProxyServiceServer will
@@ -172,6 +168,13 @@ type UnsafeProxyServiceServer interface {
 }
 
 func RegisterProxyServiceServer(s grpc.ServiceRegistrar, srv ProxyServiceServer) {
+	// If the following call pancis, it indicates UnimplementedProxyServiceServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&ProxyService_ServiceDesc, srv)
 }
 
@@ -180,42 +183,22 @@ func _ProxyService_GetTelemetry_Handler(srv interface{}, stream grpc.ServerStrea
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(ProxyServiceServer).GetTelemetry(m, &proxyServiceGetTelemetryServer{stream})
+	return srv.(ProxyServiceServer).GetTelemetry(m, &grpc.GenericServerStream[GetTelemetryRequest, GetTelemetryResponse]{ServerStream: stream})
 }
 
-type ProxyService_GetTelemetryServer interface {
-	Send(*GetTelemetryResponse) error
-	grpc.ServerStream
-}
-
-type proxyServiceGetTelemetryServer struct {
-	grpc.ServerStream
-}
-
-func (x *proxyServiceGetTelemetryServer) Send(m *GetTelemetryResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type ProxyService_GetTelemetryServer = grpc.ServerStreamingServer[GetTelemetryResponse]
 
 func _ProxyService_GetEnemyTelemetry_Handler(srv interface{}, stream grpc.ServerStream) error {
 	m := new(GetEnemyTelemetryRequest)
 	if err := stream.RecvMsg(m); err != nil {
 		return err
 	}
-	return srv.(ProxyServiceServer).GetEnemyTelemetry(m, &proxyServiceGetEnemyTelemetryServer{stream})
+	return srv.(ProxyServiceServer).GetEnemyTelemetry(m, &grpc.GenericServerStream[GetEnemyTelemetryRequest, GetEnemyTelemetryResponse]{ServerStream: stream})
 }
 
-type ProxyService_GetEnemyTelemetryServer interface {
-	Send(*GetEnemyTelemetryResponse) error
-	grpc.ServerStream
-}
-
-type proxyServiceGetEnemyTelemetryServer struct {
-	grpc.ServerStream
-}
-
-func (x *proxyServiceGetEnemyTelemetryServer) Send(m *GetEnemyTelemetryResponse) error {
-	return x.ServerStream.SendMsg(m)
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type ProxyService_GetEnemyTelemetryServer = grpc.ServerStreamingServer[GetEnemyTelemetryResponse]
 
 func _ProxyService_SendLockPacket_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(SendLockPacketRequest)
@@ -253,6 +236,24 @@ func _ProxyService_SendKamikazePacket_Handler(srv interface{}, ctx context.Conte
 	return interceptor(ctx, in, info, handler)
 }
 
+func _ProxyService_GetRedZonePacket_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(GetRedZonePacketRequest)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(ProxyServiceServer).GetRedZonePacket(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: ProxyService_GetRedZonePacket_FullMethodName,
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(ProxyServiceServer).GetRedZonePacket(ctx, req.(*GetRedZonePacketRequest))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
 // ProxyService_ServiceDesc is the grpc.ServiceDesc for ProxyService service.
 // It's only intended for direct use with grpc.RegisterService,
 // and not to be introspected or modified (even as a copy)
@@ -267,6 +268,10 @@ var ProxyService_ServiceDesc = grpc.ServiceDesc{
 		{
 			MethodName: "SendKamikazePacket",
 			Handler:    _ProxyService_SendKamikazePacket_Handler,
+		},
+		{
+			MethodName: "GetRedZonePacket",
+			Handler:    _ProxyService_GetRedZonePacket_Handler,
 		},
 	},
 	Streams: []grpc.StreamDesc{
